@@ -1,6 +1,7 @@
 [package]
 name = "z3"
-version = "0.12.1"
+rust-version = "1.85.0"
+version = "0.16.2"
 authors = ["Graydon Hoare <graydon@pobox.com>", "Bruce Mitchener <bruce.mitchener@gmail.com>", "Nick Fitzgerald <fitzgen@gmail.com>"]
 
 description = "High-level rust bindings for the Z3 SMT solver from Microsoft Research"
@@ -11,13 +12,14 @@ readme = "README.md"
 documentation = "https://docs.rs/z3/"
 homepage = "https://github.com/prove-rs/z3.rs"
 repository = "https://github.com/prove-rs/z3.rs.git"
-edition = "2018"
+edition = "2024"
 
 
 [features]
 default = []
 bundled = ["z3-sys/bundled"]
 vcpkg = ["z3-sys/vcpkg"]
+gh-release = ["z3-sys/gh-release"]
 
 # This is a legacy feature here for short term compatibility.
 static-link-z3 = ["z3-sys/bundled", "z3-sys/deprecated-static-link-z3"]
@@ -31,7 +33,9 @@ num = "0.4"
 [dev-dependencies]
 env_logger = "0.11"
 semver = "1"
+rayon = "1.10.0"
 
 [dependencies.z3-sys]
 path = "../z3-sys"
-version = "0.8"
+version = "0.9.8"
+
