@@ -2,22 +2,16 @@ use log::info;
 use std::convert::TryInto;
 use std::ops::Add;
 use std::time::Duration;
-use z3::ast::{Array, Ast, Bool, Int, BV};
+use z3::ast::{Array, Ast, BV, Bool, Int, atleast, atmost};
 use z3::*;
 
 use num::{bigint::BigInt, rational::BigRational};
 use std::str::FromStr;
+use z3::Translate;
 
 mod objectives;
 mod ops;
 mod semver_tests;
-
-fn slice_ast_eq<'ctx, T: Ast<'ctx>>(a: &[T], b: &[T]) -> bool {
-    if a.len() != b.len() {
-        return false;
-    }
-    a.iter().zip(b.iter()).all(|(a, b)| a.ast_eq(b))
-}
 
 #[test]
 fn test_config() {
@@ -27,6 +21,7 @@ fn test_config() {
 }
 
 #[test]
+#[allow(deprecated)]
 fn test_context() {
     let _ = env_logger::try_init();
     let mut cfg = Config::new();
@@ -37,49 +32,46 @@ fn test_context() {
 #[test]
 fn test_sorts_and_symbols() {
     let _ = env_logger::try_init();
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let _ = ast::Int::new_const(&ctx, "x");
-    let _ = ast::Int::new_const(&ctx, "y");
+
+    let _ = ast::Int::new_const("x");
+    let _ = ast::Int::new_const("y");
 }
 
 #[test]
 fn test_solving() {
     let _ = env_logger::try_init();
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let x = ast::Int::new_const(&ctx, "x");
-    let y = ast::Int::new_const(&ctx, "y");
 
-    let solver = Solver::new(&ctx);
-    solver.assert(&x.gt(&y));
+    let x = ast::Int::new_const("x");
+    let y = ast::Int::new_const("y");
+
+    let solver = Solver::new();
+    solver.assert(x.gt(&y));
     assert_eq!(solver.check(), SatResult::Sat);
 }
 
 #[test]
 fn test_solving_for_model() {
     let _ = env_logger::try_init();
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let x = ast::Int::new_const(&ctx, "x");
-    let y = ast::Int::new_const(&ctx, "y");
-    let zero = ast::Int::from_i64(&ctx, 0);
-    let two = ast::Int::from_i64(&ctx, 2);
-    let seven = ast::Int::from_i64(&ctx, 7);
 
-    let solver = Solver::new(&ctx);
-    solver.assert(&x.gt(&y));
-    solver.assert(&y.gt(&zero));
-    solver.assert(&y.rem(&seven).eq(&two));
-    let x_plus_two = ast::Int::add(&ctx, &[&x, &two]);
-    solver.assert(&x_plus_two.gt(&seven));
+    let x = ast::Int::new_const("x");
+    let y = ast::Int::new_const("y");
+    let zero = ast::Int::from_i64(0);
+    let two = ast::Int::from_i64(2);
+    let seven = ast::Int::from_i64(7);
+
+    let solver = Solver::new();
+    solver.assert(x.gt(&y));
+    solver.assert(y.gt(&zero));
+    solver.assert(y.rem(&seven)._eq(&two));
+    let x_plus_two = ast::Int::add(&[&x, &two]);
+    solver.assert(x_plus_two.gt(&seven));
     assert_eq!(solver.check(), SatResult::Sat);
 
     let model = solver.get_model().unwrap();
     let xv = model.eval(&x, true).unwrap().as_i64().unwrap();
     let yv = model.eval(&y, true).unwrap().as_i64().unwrap();
-    info!("x: {}", xv);
-    info!("y: {}", yv);
+    info!("x: {xv}");
+    info!("y: {yv}");
     assert!(xv > yv);
     assert!(yv % 7 == 2);
     assert!(xv + 2 > 7);
@@ -88,28 +80,27 @@ fn test_solving_for_model() {
 #[test]
 fn test_solving_for_model_cloned() {
     let _ = env_logger::try_init();
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let x = ast::Int::new_const(&ctx, "x");
-    let y = ast::Int::new_const(&ctx, "y");
-    let zero = ast::Int::from_i64(&ctx, 0);
-    let two = ast::Int::from_i64(&ctx, 2);
-    let seven = ast::Int::from_i64(&ctx, 7);
 
-    let solver = Solver::new(&ctx);
-    solver.assert(&x.gt(&y));
-    solver.assert(&y.gt(&zero));
-    solver.assert(&y.rem(&seven).eq(&two));
-    let x_plus_two = ast::Int::add(&ctx, &[&x, &two]);
-    solver.assert(&x_plus_two.gt(&seven));
+    let x = ast::Int::new_const("x");
+    let y = ast::Int::new_const("y");
+    let zero = ast::Int::from_i64(0);
+    let two = ast::Int::from_i64(2);
+    let seven = ast::Int::from_i64(7);
+
+    let solver = Solver::new();
+    solver.assert(x.gt(&y));
+    solver.assert(y.gt(&zero));
+    solver.assert(y.rem(&seven)._eq(&two));
+    let x_plus_two = ast::Int::add(&[&x, &two]);
+    solver.assert(x_plus_two.gt(&seven));
     let cloned = solver.clone();
     assert_eq!(cloned.check(), SatResult::Sat);
 
     let model = cloned.get_model().unwrap();
     let xv = model.eval(&x, true).unwrap().as_i64().unwrap();
     let yv = model.eval(&y, true).unwrap().as_i64().unwrap();
-    info!("x: {}", xv);
-    info!("y: {}", yv);
+    info!("x: {xv}");
+    info!("y: {yv}");
     assert!(xv > yv);
     assert!(yv % 7 == 2);
     assert!(xv + 2 > 7);
@@ -118,14 +109,13 @@ fn test_solving_for_model_cloned() {
 #[test]
 fn test_cloning_ast() {
     let _ = env_logger::try_init();
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let x = ast::Int::new_const(&ctx, "x");
-    let y = x.clone();
-    let zero = ast::Int::from_i64(&ctx, 0);
 
-    let solver = Solver::new(&ctx);
-    solver.assert(&x.eq(&zero));
+    let x = ast::Int::new_const("x");
+    let y = x.clone();
+    let zero = ast::Int::from_i64(0);
+
+    let solver = Solver::new();
+    solver.assert(x._eq(&zero));
     assert_eq!(solver.check(), SatResult::Sat);
 
     let model = solver.get_model().unwrap();
@@ -135,11 +125,11 @@ fn test_cloning_ast() {
     assert_eq!(yv, 0);
 }
 
-fn get_some_solver_assertions(ctx: &Context) -> Vec<ast::Bool> {
-    let s = Solver::new(ctx);
-    let x = ast::Int::new_const(ctx, "x");
-    let y = ast::Int::new_const(ctx, "y");
-    s.assert(&x.gt(&y));
+fn get_some_solver_assertions() -> Vec<ast::Bool> {
+    let s = Solver::new();
+    let x = ast::Int::new_const("x");
+    let y = ast::Int::new_const("y");
+    s.assert(x.gt(&y));
     s.get_assertions()
 }
 
@@ -149,36 +139,31 @@ fn test_solver_get_assertions_lifetime() {
     // using the solver have the lifetime of the context rather
     // than the solver.
     let _ = env_logger::try_init();
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let assertions = get_some_solver_assertions(&ctx);
+
+    let assertions = get_some_solver_assertions();
     assert_eq!(assertions.len(), 1);
 }
 
 #[test]
 fn test_format() {
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let ast = ast::Int::new_const(&ctx, "x");
+    let ast = ast::Int::new_const("x");
     assert_eq!("x", format!("{ast}"));
 
-    let int = Sort::int(&ctx);
+    let int = Sort::int();
     assert_eq!("Int", format!("{int}"));
 }
 
 #[test]
 fn test_bitvectors() {
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let a = ast::BV::new_const(&ctx, "a", 64);
-    let b = ast::BV::new_const(&ctx, "b", 64);
-    let two = ast::BV::from_i64(&ctx, 2, 64);
+    let a = ast::BV::new_const("a", 64);
+    let b = ast::BV::new_const("b", 64);
+    let two = ast::BV::from_i64(2, 64);
 
-    let solver = Solver::new(&ctx);
-    solver.assert(&a.bvsgt(&b));
-    solver.assert(&b.bvsgt(&two));
+    let solver = Solver::new();
+    solver.assert(a.bvsgt(&b));
+    solver.assert(b.bvsgt(&two));
     let b_plus_two = b.bvadd(&two);
-    solver.assert(&b_plus_two.bvsgt(&a));
+    solver.assert(b_plus_two.bvsgt(&a));
     assert_eq!(solver.check(), SatResult::Sat);
 
     let model = solver.get_model().unwrap();
@@ -191,15 +176,12 @@ fn test_bitvectors() {
 
 #[test]
 fn test_bitvector_from_str() {
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-
-    let a = ast::BV::new_const(&ctx, "a", 129);
+    let a = ast::BV::new_const("a", 129);
     // 2 ** 128
-    let b = ast::BV::from_str(&ctx, 129, "340282366920938463463374607431768211456").unwrap();
+    let b = ast::BV::from_str(129, "340282366920938463463374607431768211456").unwrap();
 
-    let solver = Solver::new(&ctx);
-    solver.assert(&a.eq(&b));
+    let solver = Solver::new();
+    solver.assert(a._eq(&b));
     assert_eq!(solver.check(), SatResult::Sat);
 
     let model = solver.get_model().unwrap();
@@ -209,13 +191,10 @@ fn test_bitvector_from_str() {
 
 #[test]
 fn test_floating_point_bits() {
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-
-    let float32 = ast::Float::new_const_float32(&ctx, "float32");
-    let float64 = ast::Float::new_const_double(&ctx, "float64");
-    let float128 = ast::Float::new_const(&ctx, "float128", 15, 113);
-    let i = ast::Int::new_const(&ctx, "int");
+    let float32 = ast::Float::new_const_float32("float32");
+    let float64 = ast::Float::new_const_double("float64");
+    let float128 = ast::Float::new_const("float128", 15, 113);
+    let i = ast::Int::new_const("int");
 
     let exp32 = Sort::float_exponent_size(&float32.get_sort());
     let sig32 = Sort::float_significand_size(&float32.get_sort());
@@ -238,25 +217,22 @@ fn test_floating_point_bits() {
 
 #[test]
 fn test_ast_translate() {
-    let cfg = Config::new();
-    let source = Context::new(&cfg);
-    let a = ast::Int::new_const(&source, "a");
+    let a = ast::Int::new_const("a");
 
-    let destination = Context::new(&cfg);
-    let translated_a = a.translate(&destination);
-
-    let slv = Solver::new(&destination);
-    slv.assert(&translated_a.eq(&ast::Int::from_u64(&destination, 2)));
-    assert_eq!(slv.check(), SatResult::Sat);
-
-    slv.assert(&translated_a.eq(&ast::Int::from_u64(&destination, 3)));
-    assert_eq!(slv.check(), SatResult::Unsat);
+    let dest_config = Config::default();
+    let sync = a.synchronized();
+    with_z3_config(&dest_config, move || {
+        let translated_a = sync.recover();
+        let slv = Solver::new();
+        slv.assert(translated_a._eq(2));
+        assert_eq!(slv.check(), SatResult::Sat);
+        slv.assert(translated_a._eq(3));
+        assert_eq!(slv.check(), SatResult::Unsat);
+    });
 }
 
 #[test]
 fn test_solver_new_from_smtlib2() {
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
     let problem = r#"
 (declare-const x Real)
 (declare-const y Real)
@@ -264,56 +240,50 @@ fn test_solver_new_from_smtlib2() {
 (assert (=( -(+(* 3 x) (* 2 y)) z) 1))
 (assert (=(+( -(* 2 x) (* 2 y)) (* 4 z)) -2))
 "#;
-    let solver = Solver::new(&ctx);
+    let solver = Solver::new();
     solver.from_string(problem);
     assert_eq!(solver.check(), SatResult::Sat);
 }
 
 #[test]
 fn test_solver_to_smtlib2() {
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let solver1 = Solver::new(&ctx);
-    let t1 = ast::Bool::from_bool(&ctx, true);
-    let t2 = ast::Bool::from_bool(&ctx, true);
-    solver1.assert(&t1.eq(&t2));
+    let solver1 = Solver::new();
+    let t1 = ast::Bool::from_bool(true);
+    let t2 = ast::Bool::from_bool(true);
+    solver1.assert(t1._eq(&t2));
     let s1_smt2 = solver1.to_smt2();
-    let solver2 = Solver::new(&ctx);
+    let solver2 = Solver::new();
     solver2.from_string(s1_smt2);
     assert_eq!(solver2.check(), solver1.check());
 }
 
 #[test]
 fn test_solver_translate() {
-    let cfg = Config::new();
-    let source = Context::new(&cfg);
-    let a = ast::Int::new_const(&source, "a");
-
-    let destination = Context::new(&cfg);
-    let translated_a = a.translate(&destination);
-
-    let slv = Solver::new(&destination);
-    slv.assert(&translated_a.eq(&ast::Int::from_u64(&destination, 2)));
-    assert_eq!(slv.check(), SatResult::Sat);
-
-    let translated_slv = slv.translate(&source);
-    // Add a new constraint, make the old one unsatisfiable, while the copy remains satisfiable.
-    slv.assert(&translated_a.eq(&ast::Int::from_u64(&destination, 3)));
+    let a = ast::Int::new_const("a");
+    let s = a.synchronized();
+    let slv = with_z3_config(&Config::new(), || {
+        let a = s.recover();
+        let slv = Solver::new();
+        slv.assert(a._eq(2));
+        assert_eq!(slv.check(), SatResult::Sat);
+        slv.synchronized()
+    })
+    .recover();
+    slv.assert(a._eq(3));
     assert_eq!(slv.check(), SatResult::Unsat);
-    assert_eq!(translated_slv.check(), SatResult::Sat);
 }
 
 #[test]
+#[allow(deprecated)]
 fn test_model_translate() {
     let cfg = Config::new();
-    let source = Context::new(&cfg);
-    let a = ast::Int::new_const(&source, "a");
+    let a = ast::Int::new_const("a");
 
     let destination = Context::new(&cfg);
     let translated_a = a.translate(&destination);
 
-    let slv = Solver::new(&source);
-    slv.assert(&a.eq(&ast::Int::from_u64(&source, 2)));
+    let slv = Solver::new();
+    slv.assert(a._eq(2));
     assert_eq!(slv.check(), SatResult::Sat);
 
     let model = slv.get_model().unwrap();
@@ -331,54 +301,50 @@ fn test_model_translate() {
 
 #[test]
 fn test_pb_ops_model() {
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let x = ast::Bool::new_const(&ctx, "x");
-    let y = ast::Bool::new_const(&ctx, "y");
-    let solver = Solver::new(&ctx);
+    let x = ast::Bool::new_const("x");
+    let y = ast::Bool::new_const("y");
+    let solver = Solver::new();
     solver.push();
 
-    solver.assert(&ast::Bool::pb_eq(&ctx, &[(&x, 1), (&y, 1)], 1));
+    solver.assert(ast::Bool::pb_eq(&[(&x, 1), (&y, 1)], 1));
     assert_eq!(solver.check(), SatResult::Sat);
     let model = solver.get_model().unwrap();
     let xv = model.eval(&x, true).unwrap().as_bool().unwrap();
     let yv = model.eval(&y, true).unwrap().as_bool().unwrap();
-    info!("x: {}", xv);
-    info!("y: {}", yv);
+    info!("x: {xv}");
+    info!("y: {yv}");
     assert!((xv && !yv) || (!xv && yv));
 
     solver.pop(1);
     solver.push();
-    solver.assert(&ast::Bool::pb_ge(&ctx, &[(&x, 1), (&y, 1)], 2));
+    solver.assert(ast::Bool::pb_ge(&[(&x, 1), (&y, 1)], 2));
     assert_eq!(solver.check(), SatResult::Sat);
     let model = solver.get_model().unwrap();
     let xv = model.eval(&x, true).unwrap().as_bool().unwrap();
     let yv = model.eval(&y, true).unwrap().as_bool().unwrap();
-    info!("x: {}", xv);
-    info!("y: {}", yv);
+    info!("x: {xv}");
+    info!("y: {yv}");
     assert!(xv && yv);
 
     solver.pop(1);
-    solver.assert(&ast::Bool::pb_le(&ctx, &[(&x, 1), (&y, 1)], 0));
+    solver.assert(ast::Bool::pb_le(&[(&x, 1), (&y, 1)], 0));
     assert_eq!(solver.check(), SatResult::Sat);
     let model = solver.get_model().unwrap();
     let xv = model.eval(&x, true).unwrap().as_bool().unwrap();
     let yv = model.eval(&y, true).unwrap().as_bool().unwrap();
-    info!("x: {}", xv);
-    info!("y: {}", yv);
+    info!("x: {xv}");
+    info!("y: {yv}");
     assert!(!xv && !yv);
 }
 
 #[test]
 fn function_ref_count() {
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let solver = Solver::new(&ctx);
+    let solver = Solver::new();
 
-    let int_sort = Sort::int(&ctx);
+    let int_sort = Sort::int();
 
-    let _f = FuncDecl::new(&ctx, "f", &[&int_sort], &int_sort);
-    let _g = FuncDecl::new(&ctx, "g", &[&int_sort], &int_sort);
+    let _f = FuncDecl::new("f", &[&int_sort], &int_sort);
+    let _g = FuncDecl::new("g", &[&int_sort], &int_sort);
 
     assert_eq!(solver.check(), SatResult::Sat);
 }
@@ -386,19 +352,18 @@ fn function_ref_count() {
 #[test]
 fn test_params() {
     let _ = env_logger::try_init();
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let x = ast::Int::new_const(&ctx, "x");
-    let y = ast::Int::new_const(&ctx, "y");
 
-    let mut params = Params::new(&ctx);
+    let x = ast::Int::new_const("x");
+    let y = ast::Int::new_const("y");
+
+    let mut params = Params::new();
     params.set_bool("smt.mbqi", false);
     params.set_f64("smt.qi.eager_threshold", 5.0);
     params.set_u32("smt.qi.max_instances", 999);
 
-    let solver = Solver::new(&ctx);
+    let solver = Solver::new();
     solver.set_params(&params);
-    solver.assert(&x.gt(&y));
+    solver.assert(x.gt(&y));
     assert_eq!(solver.check(), SatResult::Sat);
 }
 
@@ -421,31 +386,26 @@ fn test_global_params() {
 
 #[test]
 fn test_substitution() {
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
+    let x = ast::Real::new_const("x");
+    let y = ast::Real::new_const("y");
+    let z = ast::Real::new_const("z");
 
-    let x = ast::Real::new_const(&ctx, "x");
-    let y = ast::Real::new_const(&ctx, "y");
-    let z = ast::Real::new_const(&ctx, "z");
-
-    let x_plus_y = ast::Real::add(&ctx, &[&x, &y]);
-    let x_plus_z = ast::Real::add(&ctx, &[&x, &z]);
+    let x_plus_y = ast::Real::add(&[&x, &y]);
+    let x_plus_z = ast::Real::add(&[&x, &z]);
 
     let substitutions = &[(&y, &z)];
 
-    assert!(x_plus_y.substitute(substitutions).ast_eq(&x_plus_z));
+    assert!(x_plus_y.substitute(substitutions) == x_plus_z);
 }
 
 #[test]
 fn test_real_cmp() {
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let solver = Solver::new(&ctx);
+    let solver = Solver::new();
 
-    let x = ast::Real::new_const(&ctx, "x");
-    let x_plus_1 = ast::Real::add(&ctx, &[&x, &ast::Real::from_rational(&ctx, 1, 1)]);
+    let x = ast::Real::new_const("x");
+    let x_plus_1 = ast::Real::add(&[&x, &ast::Real::from_real(1, 1)]);
     // forall x, x < x + 1
-    let forall = ast::forall_const(&ctx, &[&x], &[], &x.lt(&x_plus_1));
+    let forall = ast::forall_const(&[&x], &[], &x.lt(&x_plus_1));
 
     solver.assert(&forall);
     assert_eq!(solver.check(), SatResult::Sat);
@@ -453,155 +413,118 @@ fn test_real_cmp() {
 
 #[test]
 fn test_float() {
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-
-    let f = ast::Float::from_f64(&ctx, 1.0);
+    let f = ast::Float::from_f64(1.0);
     assert_eq!(f.as_f64(), 1.0);
 }
 
 #[test]
 fn test_float_add() {
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let solver = Solver::new(&ctx);
+    let solver = Solver::new();
 
-    let x = ast::Float::new_const_float32(&ctx, "x");
-    let x_plus_one = ast::Float::round_towards_zero(&ctx).add(&x, &ast::Float::from_f32(&ctx, 1.0));
-    let y = ast::Float::from_f32(&ctx, std::f32::consts::PI);
+    let x = ast::Float::new_const_float32("x");
+    let x_plus_one = ast::RoundingMode::round_towards_zero().add(&x, 1.0f32);
+    let y = ast::Float::from_f32(std::f32::consts::PI);
 
-    solver.assert(&x_plus_one.eq(&y));
+    solver.assert(x_plus_one._eq(y));
     assert_eq!(solver.check(), SatResult::Sat);
 }
 
 #[test]
 fn test_arbitrary_size_real() {
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let solver = Solver::new(&ctx);
+    let solver = Solver::new();
 
-    let x =
-        ast::Real::from_rational_str(&ctx, "99999999999999999999998", "99999999999999999999999")
-            .unwrap();
-    let y = ast::Real::from_rational(&ctx, 1, 1);
+    let x = ast::Real::from_real_str("99999999999999999999998", "99999999999999999999999").unwrap();
+    let y = ast::Real::from_real(1, 1);
 
-    solver.assert(&x.lt(&y));
+    solver.assert(x.lt(&y));
     assert_eq!(solver.check(), SatResult::Sat);
 }
 
 #[test]
 fn test_arbitrary_size_int() {
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let solver = Solver::new(&ctx);
+    let solver = Solver::new();
 
-    let x = ast::Int::from_str(&ctx, "99999999999999999999998").unwrap();
-    let one = ast::Int::from_i64(&ctx, 1);
-    let y = ast::Int::from_str(&ctx, "99999999999999999999999").unwrap();
+    let x = ast::Int::from_str("99999999999999999999998").unwrap();
+    let one = ast::Int::from_i64(1);
+    let y = ast::Int::from_str("99999999999999999999999").unwrap();
 
-    solver.assert(&ast::Int::add(&ctx, &[&x, &one]).eq(&y));
+    solver.assert(ast::Int::add(&[&x, &one])._eq(&y));
     assert_eq!(solver.check(), SatResult::Sat);
 }
 
 #[test]
 fn test_arbitrary_size_real_from_bigrational() {
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let solver = Solver::new(&ctx);
+    let solver = Solver::new();
 
-    let x =
-        ast::Real::from_rational_str(&ctx, "99999999999999999999998", "99999999999999999999999")
-            .unwrap();
+    let x = ast::Real::from_real_str("99999999999999999999998", "99999999999999999999999").unwrap();
     let num = BigInt::from_str("99999999999999999999998").unwrap();
     let den = BigInt::from_str("99999999999999999999999").unwrap();
     let ratio = BigRational::new(num, den);
-    let y = ast::Real::from_big_rational(&ctx, &ratio);
 
-    solver.assert(&x.eq(&y));
+    solver.assert(x._eq(ratio));
     assert_eq!(solver.check(), SatResult::Sat);
 }
 
 #[test]
 fn test_arbitrary_size_int_from_bigint() {
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let solver = Solver::new(&ctx);
+    let solver = Solver::new();
 
     let num1 = BigInt::from_str("99999999999999999999998").unwrap();
-    let x = ast::Int::from_big_int(&ctx, &num1);
-    let y = ast::Int::from_i64(&ctx, 1);
+    let x = ast::Int::from_big_int(&num1);
+    let y = ast::Int::from_i64(1);
 
     let num2 = BigInt::from_str("99999999999999999999999").unwrap();
-    let z = ast::Int::from_big_int(&ctx, &num2);
 
-    solver.assert(&ast::Int::add(&ctx, &[&x, &y]).eq(&z));
+    solver.assert(ast::Int::add(&[&x, &y])._eq(num2));
     assert_eq!(solver.check(), SatResult::Sat);
 }
 
 #[test]
 fn test_string_eq() {
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let solver = Solver::new(&ctx);
+    let solver = Solver::new();
 
-    let x = ast::String::from_str(&ctx, "foo").unwrap();
-    let y = ast::String::from_str(&ctx, "foo").unwrap();
-    let z = ast::String::from_str(&ctx, "bar").unwrap();
-    let h = ast::String::new_const(&ctx, "h");
+    let x = ast::String::from_str("foo").unwrap();
+    let h = ast::String::new_const("h");
 
-    solver.assert(&x.eq(&y));
-    solver.assert(&x.eq(&z).not());
-    solver.assert(&h.eq(&x));
+    solver.assert(x._eq("foo"));
+    solver.assert(x._eq("bar").not());
+    solver.assert(h._eq(&x));
     assert_eq!(solver.check(), SatResult::Sat);
 
-    solver.assert(&h.eq(&z));
+    solver.assert(h._eq("bar"));
     assert_eq!(solver.check(), SatResult::Unsat);
 }
 
 #[test]
 fn test_string_concat() {
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let solver = Solver::new(&ctx);
+    let solver = Solver::new();
 
-    let x = ast::String::from_str(&ctx, "foo").unwrap();
-    let y = ast::String::from_str(&ctx, "bar").unwrap();
-    let z = ast::String::from_str(&ctx, "foobar").unwrap();
-
-    solver.assert(&ast::String::concat(&ctx, &[&x, &y]).eq(&z));
+    solver.assert(ast::String::concat(&["foo", "bar"])._eq("foobar"));
     assert_eq!(solver.check(), SatResult::Sat);
 }
 
 #[test]
 fn test_string_prefix() {
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let solver = Solver::new(&ctx);
+    let solver = Solver::new();
 
-    let x = ast::String::from_str(&ctx, "foo").unwrap();
-    let y = ast::String::from_str(&ctx, "foobar").unwrap();
+    let x = ast::String::from_str("foo").unwrap();
 
-    solver.assert(&x.prefix(&y));
+    solver.assert(x.prefix("foobar"));
     assert_eq!(solver.check(), SatResult::Sat);
 }
 
 #[test]
 fn test_string_suffix() {
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let solver = Solver::new(&ctx);
+    let solver = Solver::new();
 
-    let x = ast::String::from_str(&ctx, "bar").unwrap();
-    let y = ast::String::from_str(&ctx, "foobar").unwrap();
+    let x = ast::String::from_str("bar").unwrap();
 
-    solver.assert(&x.suffix(&y));
+    solver.assert(x.suffix("foobar"));
     assert_eq!(solver.check(), SatResult::Sat);
 }
 
 fn assert_string_roundtrip(source: &str) {
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let expr = ast::String::from_str(&ctx, source).unwrap();
+    let expr = ast::String::from_str(source).unwrap();
     assert_eq!(&expr.as_string().unwrap(), source);
 }
 
@@ -616,31 +539,28 @@ fn test_string_as_string() {
 #[test]
 fn test_rec_func_def() {
     let _ = env_logger::try_init();
-    let cfg = Config::new();
 
-    let ctx = Context::new(&cfg);
-
-    let fac = RecFuncDecl::new(&ctx, "fac", &[&Sort::int(&ctx)], &Sort::int(&ctx));
-    let n = ast::Int::new_const(&ctx, "n");
-    let n_minus_1 = ast::Int::sub(&ctx, &[&n, &ast::Int::from_i64(&ctx, 1)]);
+    let fac = RecFuncDecl::new("fac", &[&Sort::int()], &Sort::int());
+    let n = ast::Int::new_const("n");
+    let n_minus_1 = &n - 1;
     let fac_of_n_minus_1 = fac.apply(&[&n_minus_1]);
-    let cond: ast::Bool = n.le(&ast::Int::from_i64(&ctx, 0));
+    let cond: ast::Bool = n.le(0);
     let body = cond.ite(
-        &ast::Int::from_i64(&ctx, 1),
-        &ast::Int::mul(&ctx, &[&n, &fac_of_n_minus_1.as_int().unwrap()]),
+        &ast::Int::from_i64(1),
+        &ast::Int::mul(&[&n, &fac_of_n_minus_1.as_int().unwrap()]),
     );
 
     fac.add_def(&[&n], &body);
 
-    let x = ast::Int::new_const(&ctx, "x");
-    let y = ast::Int::new_const(&ctx, "y");
+    let x = ast::Int::new_const("x");
+    let y = ast::Int::new_const("y");
 
-    let solver = Solver::new(&ctx);
+    let solver = Solver::new();
 
-    solver.assert(&x.eq(&fac.apply(&[&ast::Int::from_i64(&ctx, 4)]).as_int().unwrap()));
-    solver.assert(&y.eq(&ast::Int::mul(&ctx, &[&ast::Int::from_i64(&ctx, 5), &x])));
-    solver.assert(&y.eq(&fac.apply(&[&ast::Int::from_i64(&ctx, 5)]).as_int().unwrap()));
-    solver.assert(&y.eq(&ast::Int::from_i64(&ctx, 120)));
+    solver.assert(x._eq(fac.apply(&[&ast::Int::from_i64(4)]).as_int().unwrap()));
+    solver.assert(y._eq(ast::Int::mul(&[&ast::Int::from_i64(5), &x])));
+    solver.assert(y._eq(fac.apply(&[&ast::Int::from_i64(5)]).as_int().unwrap()));
+    solver.assert(y._eq(120));
 
     assert_eq!(solver.check(), SatResult::Sat);
 }
@@ -648,34 +568,31 @@ fn test_rec_func_def() {
 #[test]
 fn test_rec_func_def_unsat() {
     let _ = env_logger::try_init();
-    let cfg = Config::new();
 
-    let ctx = Context::new(&cfg);
-
-    let fac = RecFuncDecl::new(&ctx, "fac", &[&Sort::int(&ctx)], &Sort::int(&ctx));
-    let n = ast::Int::new_const(&ctx, "n");
-    let n_minus_1 = ast::Int::sub(&ctx, &[&n, &ast::Int::from_i64(&ctx, 1)]);
+    let fac = RecFuncDecl::new("fac", &[&Sort::int()], &Sort::int());
+    let n = ast::Int::new_const("n");
+    let n_minus_1 = ast::Int::sub(&[&n, &ast::Int::from_i64(1)]);
     let fac_of_n_minus_1 = fac.apply(&[&n_minus_1]);
-    let cond: ast::Bool = n.le(&ast::Int::from_i64(&ctx, 0));
+    let cond: ast::Bool = n.le(0);
     let body = cond.ite(
-        &ast::Int::from_i64(&ctx, 1),
-        &ast::Int::mul(&ctx, &[&n, &fac_of_n_minus_1.as_int().unwrap()]),
+        &ast::Int::from_i64(1),
+        &ast::Int::mul(&[&n, &fac_of_n_minus_1.as_int().unwrap()]),
     );
 
     fac.add_def(&[&n], &body);
 
-    let x = ast::Int::new_const(&ctx, "x");
-    let y = ast::Int::new_const(&ctx, "y");
+    let x = ast::Int::new_const("x");
+    let y = ast::Int::new_const("y");
 
-    let solver = Solver::new(&ctx);
+    let solver = Solver::new();
 
-    solver.assert(&x.eq(&fac.apply(&[&ast::Int::from_i64(&ctx, 4)]).as_int().unwrap()));
-    solver.assert(&y.eq(&ast::Int::mul(&ctx, &[&ast::Int::from_i64(&ctx, 5), &x])));
-    solver.assert(&y.eq(&fac.apply(&[&ast::Int::from_i64(&ctx, 5)]).as_int().unwrap()));
+    solver.assert(x._eq(fac.apply(&[&ast::Int::from_i64(4)]).as_int().unwrap()));
+    solver.assert(y._eq(ast::Int::mul(&[&ast::Int::from_i64(5), &x])));
+    solver.assert(y._eq(fac.apply(&[&ast::Int::from_i64(5)]).as_int().unwrap()));
 
     // If fac was an uninterpreted function, this assertion would work.
     // To see this, comment out `fac.add_def(&[&n.into()], &body);`
-    solver.assert(&y.eq(&ast::Int::from_i64(&ctx, 25)));
+    solver.assert(y._eq(25));
 
     assert_eq!(solver.check(), SatResult::Unsat);
 }
@@ -686,24 +603,24 @@ fn test_solver_unknown() {
     let mut cfg = Config::new();
     // Use a very short timeout to quickly return "unknown"
     cfg.set_timeout_msec(1);
-    let ctx = Context::new(&cfg);
+    with_z3_config(&cfg, || {
+        // An open problem: find a model for x^3 + y^3 + z^3 == 42
+        // See: https://en.wikipedia.org/wiki/Sums_of_three_cubes
+        let x = ast::Int::new_const("x");
+        let y = ast::Int::new_const("y");
+        let z = ast::Int::new_const("z");
+        let x_cube = ast::Int::mul(&[&x, &x, &x]);
+        let y_cube = ast::Int::mul(&[&y, &y, &y]);
+        let z_cube = ast::Int::mul(&[&z, &z, &z]);
+        let sum_of_cubes = x_cube + y_cube + z_cube;
+        let sum_of_cubes_is_42 = sum_of_cubes._eq(42);
 
-    // An open problem: find a model for x^3 + y^3 + z^3 == 42
-    // See: https://en.wikipedia.org/wiki/Sums_of_three_cubes
-    let x = ast::Int::new_const(&ctx, "x");
-    let y = ast::Int::new_const(&ctx, "y");
-    let z = ast::Int::new_const(&ctx, "z");
-    let x_cube = ast::Int::mul(&ctx, &[&x, &x, &x]);
-    let y_cube = ast::Int::mul(&ctx, &[&y, &y, &y]);
-    let z_cube = ast::Int::mul(&ctx, &[&z, &z, &z]);
-    let sum_of_cubes = &ast::Int::add(&ctx, &[&x_cube, &y_cube, &z_cube]);
-    let sum_of_cubes_is_42 = sum_of_cubes.eq(&ast::Int::from_i64(&ctx, 42));
+        let solver = Solver::new();
+        solver.assert(&sum_of_cubes_is_42);
 
-    let solver = Solver::new(&ctx);
-    solver.assert(&sum_of_cubes_is_42);
-
-    assert_eq!(solver.check(), SatResult::Unknown);
-    assert!(solver.get_reason_unknown().is_some());
+        assert_eq!(solver.check(), SatResult::Unknown);
+        assert!(solver.get_reason_unknown().is_some());
+    });
 }
 
 #[test]
@@ -712,31 +629,31 @@ fn test_optimize_unknown() {
     let mut cfg = Config::new();
     // Use a very short timeout to quickly return "unknown"
     cfg.set_timeout_msec(1);
-    let ctx = Context::new(&cfg);
 
-    // An open problem: find a model for x^3 + y^3 + z^3 == 42
-    // See: https://en.wikipedia.org/wiki/Sums_of_three_cubes
-    let x = ast::Int::new_const(&ctx, "x");
-    let y = ast::Int::new_const(&ctx, "y");
-    let z = ast::Int::new_const(&ctx, "z");
-    let x_cube = ast::Int::mul(&ctx, &[&x, &x, &x]);
-    let y_cube = ast::Int::mul(&ctx, &[&y, &y, &y]);
-    let z_cube = ast::Int::mul(&ctx, &[&z, &z, &z]);
-    let sum_of_cubes = &ast::Int::add(&ctx, &[&x_cube, &y_cube, &z_cube]);
-    let sum_of_cubes_is_42 = sum_of_cubes.eq(&ast::Int::from_i64(&ctx, 42));
+    with_z3_config(&cfg, || {
+        // An open problem: find a model for x^3 + y^3 + z^3 == 42
+        // See: https://en.wikipedia.org/wiki/Sums_of_three_cubes
+        let x = ast::Int::new_const("x");
+        let y = ast::Int::new_const("y");
+        let z = ast::Int::new_const("z");
+        let x_cube = ast::Int::mul(&[&x, &x, &x]);
+        let y_cube = ast::Int::mul(&[&y, &y, &y]);
+        let z_cube = ast::Int::mul(&[&z, &z, &z]);
+        let sum_of_cubes = x_cube + y_cube + z_cube;
+        let sum_of_cubes_is_42 = sum_of_cubes._eq(42);
 
-    let optimize = Optimize::new(&ctx);
-    optimize.assert(&sum_of_cubes_is_42);
+        let optimize = Optimize::new();
+        optimize.assert(&sum_of_cubes_is_42);
 
-    assert_eq!(optimize.check(&[]), SatResult::Unknown);
-    assert!(optimize.get_reason_unknown().is_some());
+        assert_eq!(optimize.check(&[]), SatResult::Unknown);
+        assert!(optimize.get_reason_unknown().is_some());
+    });
 }
 
 #[test]
 fn test_optimize_new_from_smtlib2() {
     let _ = env_logger::try_init();
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
+
     let problem = r#"
 (declare-const x Real)
 (declare-const y Real)
@@ -744,7 +661,7 @@ fn test_optimize_new_from_smtlib2() {
 (assert (=( -(+(* 3 x) (* 2 y)) z) 1))
 (assert (=(+( -(* 2 x) (* 2 y)) (* 4 z)) -2))
 "#;
-    let optimize = Optimize::new(&ctx);
+    let optimize = Optimize::new();
     optimize.from_string(problem);
     assert_eq!(optimize.check(&[]), SatResult::Sat);
 }
@@ -753,22 +670,20 @@ fn test_optimize_new_from_smtlib2() {
 fn test_get_unsat_core() {
     let _ = env_logger::try_init();
 
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let solver = Solver::new(&ctx);
+    let solver = Solver::new();
 
     assert!(
         solver.get_unsat_core().is_empty(),
         "no unsat core before assertions"
     );
 
-    let x = ast::Int::new_const(&ctx, "x");
+    let x = ast::Int::new_const("x");
 
-    let x_is_three = ast::Bool::new_const(&ctx, "x-is-three");
-    solver.assert_and_track(&x.eq(&ast::Int::from_i64(&ctx, 3)), &x_is_three);
+    let x_is_three = ast::Bool::new_const("x-is-three");
+    solver.assert_and_track(x._eq(3), &x_is_three);
 
-    let x_is_five = ast::Bool::new_const(&ctx, "x-is-five");
-    solver.assert_and_track(&x.eq(&ast::Int::from_i64(&ctx, 5)), &x_is_five);
+    let x_is_five = ast::Bool::new_const("x-is-five");
+    solver.assert_and_track(x._eq(5), &x_is_five);
 
     assert!(
         solver.get_unsat_core().is_empty(),
@@ -780,30 +695,28 @@ fn test_get_unsat_core() {
 
     let unsat_core = solver.get_unsat_core();
     assert_eq!(unsat_core.len(), 2);
-    assert!(unsat_core.iter().any(|x| x.ast_eq(&x_is_three)));
-    assert!(unsat_core.iter().any(|x| x.ast_eq(&x_is_five)));
+    assert!(unsat_core.contains(&x_is_three));
+    assert!(unsat_core.contains(&x_is_five));
 }
 
 #[test]
 fn test_optimize_get_unsat_core() {
     let _ = env_logger::try_init();
 
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let optimize = Optimize::new(&ctx);
+    let optimize = Optimize::new();
 
     assert!(
         optimize.get_unsat_core().is_empty(),
         "no unsat core before assertions"
     );
 
-    let x = Int::new_const(&ctx, "x");
+    let x = Int::new_const("x");
 
-    let x_is_three = Bool::new_const(&ctx, "x-is-three");
-    optimize.assert_and_track(&x.eq(&Int::from_i64(&ctx, 3)), &x_is_three);
+    let x_is_three = Bool::new_const("x-is-three");
+    optimize.assert_and_track(&x._eq(3), &x_is_three);
 
-    let x_is_five = Bool::new_const(&ctx, "x-is-five");
-    optimize.assert_and_track(&x.eq(&Int::from_i64(&ctx, 5)), &x_is_five);
+    let x_is_five = Bool::new_const("x-is-five");
+    optimize.assert_and_track(&x._eq(5), &x_is_five);
 
     assert!(
         optimize.get_unsat_core().is_empty(),
@@ -815,40 +728,35 @@ fn test_optimize_get_unsat_core() {
 
     let unsat_core = optimize.get_unsat_core();
     assert_eq!(unsat_core.len(), 2);
-    assert!(unsat_core.iter().any(|x| x.ast_eq(&x_is_three)));
-    assert!(unsat_core.iter().any(|x| x.ast_eq(&x_is_five)));
+    assert!(unsat_core.contains(&x_is_three));
+    assert!(unsat_core.contains(&x_is_five));
 
     // try check API
 
-    let a = x.eq(&Int::from_i64(&ctx, 4));
-    let b = x.eq(&Int::from_i64(&ctx, 6));
+    let a = x._eq(4);
+    let b = x._eq(6);
     let result = optimize.check(&[a.clone(), b.clone()]);
     assert_eq!(result, SatResult::Unsat);
 
     let unsat_core = optimize.get_unsat_core();
     assert_eq!(unsat_core.len(), 2);
-    assert!(unsat_core.iter().any(|x| x.ast_eq(&a)));
-    assert!(unsat_core.iter().any(|x| x.ast_eq(&b)));
+    assert!(unsat_core.contains(&a));
+    assert!(unsat_core.contains(&b));
 }
 
 #[test]
 fn test_datatype_builder() {
     let _ = env_logger::try_init();
 
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let solver = Solver::new(&ctx);
+    let solver = Solver::new();
 
-    let maybe_int = DatatypeBuilder::new(&ctx, "MaybeInt")
+    let maybe_int = DatatypeBuilder::new("MaybeInt")
         .variant("Nothing", vec![])
-        .variant(
-            "Just",
-            vec![("int", DatatypeAccessor::Sort(Sort::int(&ctx)))],
-        )
+        .variant("Just", vec![("int", DatatypeAccessor::Sort(Sort::int()))])
         .finish();
 
     let nothing = maybe_int.variants[0].constructor.apply(&[]);
-    let five = ast::Int::from_i64(&ctx, 5);
+    let five = ast::Int::from_i64(5);
     let just_five = maybe_int.variants[1].constructor.apply(&[&five]);
 
     let nothing_is_nothing = maybe_int.variants[0]
@@ -863,14 +771,14 @@ fn test_datatype_builder() {
         .apply(&[&nothing])
         .as_bool()
         .unwrap();
-    solver.assert(&nothing_is_just.not());
+    solver.assert(nothing_is_just.not());
 
     let just_five_is_nothing = maybe_int.variants[0]
         .tester
         .apply(&[&just_five])
         .as_bool()
         .unwrap();
-    solver.assert(&just_five_is_nothing.not());
+    solver.assert(just_five_is_nothing.not());
 
     let just_five_is_just = maybe_int.variants[1]
         .tester
@@ -883,7 +791,7 @@ fn test_datatype_builder() {
         .apply(&[&just_five])
         .as_int()
         .unwrap();
-    solver.assert(&five.eq(&five_two));
+    solver.assert(five._eq(&five_two));
 
     assert_eq!(solver.check(), SatResult::Sat);
 }
@@ -891,16 +799,15 @@ fn test_datatype_builder() {
 #[test]
 fn test_recursive_datatype() {
     let _ = env_logger::try_init();
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let solver = Solver::new(&ctx);
 
-    let list_sort = DatatypeBuilder::new(&ctx, "List")
+    let solver = Solver::new();
+
+    let list_sort = DatatypeBuilder::new("List")
         .variant("nil", vec![])
         .variant(
             "cons",
             vec![
-                ("car", DatatypeAccessor::Sort(Sort::int(&ctx))),
+                ("car", DatatypeAccessor::Sort(Sort::int())),
                 ("cdr", DatatypeAccessor::Datatype("List".into())),
             ],
         )
@@ -908,7 +815,7 @@ fn test_recursive_datatype() {
 
     assert_eq!(list_sort.variants.len(), 2);
     let nil = list_sort.variants[0].constructor.apply(&[]);
-    let five = ast::Int::from_i64(&ctx, 5);
+    let five = ast::Int::from_i64(5);
     let cons_five_nil = list_sort.variants[1].constructor.apply(&[&five, &nil]);
 
     let nil_is_nil = list_sort.variants[0]
@@ -923,14 +830,14 @@ fn test_recursive_datatype() {
         .apply(&[&nil])
         .as_bool()
         .unwrap();
-    solver.assert(&nil_is_cons.not());
+    solver.assert(nil_is_cons.not());
 
     let cons_five_nil_is_nil = list_sort.variants[0]
         .tester
         .apply(&[&cons_five_nil])
         .as_bool()
         .unwrap();
-    solver.assert(&cons_five_nil_is_nil.not());
+    solver.assert(cons_five_nil_is_nil.not());
 
     let cons_five_nil_is_cons = list_sort.variants[1]
         .tester
@@ -943,35 +850,31 @@ fn test_recursive_datatype() {
         .apply(&[&cons_five_nil])
         .as_int()
         .unwrap();
-    solver.assert(&car_cons_five_is_five.eq(&five));
+    solver.assert(car_cons_five_is_five._eq(&five));
     assert_eq!(solver.check(), SatResult::Sat);
 
     let cdr_cons_five_is_nil = list_sort.variants[1].accessors[1]
         .apply(&[&cons_five_nil])
         .as_datatype()
         .unwrap();
-    solver.assert(&cdr_cons_five_is_nil.eq(&nil.as_datatype().unwrap()));
+    solver.assert(cdr_cons_five_is_nil._eq(nil.as_datatype().unwrap()));
     assert_eq!(solver.check(), SatResult::Sat);
 }
 
 #[test]
 fn test_mutually_recursive_datatype() {
     let _ = env_logger::try_init();
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let solver = Solver::new(&ctx);
 
-    let tree_builder = DatatypeBuilder::new(&ctx, "Tree")
-        .variant(
-            "leaf",
-            vec![("val", DatatypeAccessor::Sort(Sort::int(&ctx)))],
-        )
+    let solver = Solver::new();
+
+    let tree_builder = DatatypeBuilder::new("Tree")
+        .variant("leaf", vec![("val", DatatypeAccessor::Sort(Sort::int()))])
         .variant(
             "node",
             vec![("children", DatatypeAccessor::Datatype("TreeList".into()))],
         );
 
-    let tree_list_builder = DatatypeBuilder::new(&ctx, "TreeList")
+    let tree_list_builder = DatatypeBuilder::new("TreeList")
         .variant("nil", vec![])
         .variant(
             "cons",
@@ -993,17 +896,17 @@ fn test_mutually_recursive_datatype() {
     assert_eq!(tree_list_sort.variants[0].accessors.len(), 0);
     assert_eq!(tree_list_sort.variants[1].accessors.len(), 2);
 
-    let ten = ast::Int::from_i64(&ctx, 10);
+    let ten = ast::Int::from_i64(10);
     let leaf_ten = tree_sort.variants[0].constructor.apply(&[&ten]);
     let leaf_ten_val_is_ten = tree_sort.variants[0].accessors[0]
         .apply(&[&leaf_ten])
         .as_int()
         .unwrap();
-    solver.assert(&leaf_ten_val_is_ten.eq(&ten.clone()));
+    solver.assert(leaf_ten_val_is_ten._eq(ten.clone()));
     assert_eq!(solver.check(), SatResult::Sat);
 
     let nil = tree_list_sort.variants[0].constructor.apply(&[]);
-    let twenty = ast::Int::from_i64(&ctx, 20);
+    let twenty = ast::Int::from_i64(20);
     let leaf_twenty = tree_sort.variants[0].constructor.apply(&[&twenty]);
     let cons_leaf_twenty_nil = tree_list_sort.variants[1]
         .constructor
@@ -1021,42 +924,38 @@ fn test_mutually_recursive_datatype() {
     // n2 = Tree.node(TreeList.cons(n1, TreeList.nil))
     let n2 = tree_sort.variants[1].constructor.apply(&[&n1_cons_nil]);
 
-    solver.assert(&n2.eq(&n1).not());
+    solver.assert(n2._eq(&n1).not());
 
     // assert(TreeList.car(Tree.children(n2)) == n1)
     solver.assert(
-        &tree_list_sort.variants[1].accessors[0]
+        tree_list_sort.variants[1].accessors[0]
             .apply(&[&tree_sort.variants[1].accessors[0].apply(&[&n2])])
-            .eq(&n1),
+            ._eq(&n1),
     );
     assert_eq!(solver.check(), SatResult::Sat);
 }
 
 #[test]
 fn get_model_without_check_does_not_exit() {
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let solver = Solver::new(&ctx);
+    let solver = Solver::new();
     solver.get_model();
 }
 
 #[test]
 fn check_application_of_tactic_to_goal() {
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let params = Params::new(&ctx);
+    let params = Params::new();
 
-    let tactic = Tactic::new(&ctx, "simplify");
-    let repeat_tactic = Tactic::repeat(&ctx, &tactic, 100);
+    let tactic = Tactic::new("simplify");
+    let repeat_tactic = Tactic::repeat(&tactic, 100);
 
-    let goal = Goal::new(&ctx, false, false, false);
-    let x = ast::Bool::new_const(&ctx, "x");
+    let goal = Goal::new(false, false, false);
+    let x = ast::Bool::new_const("x");
     goal.assert(&x);
 
-    let one = ast::Int::from_i64(&ctx, 1);
-    let two = ast::Int::from_i64(&ctx, 2);
+    let one = ast::Int::from_i64(1);
+    let two = ast::Int::from_i64(2);
 
-    let y = ast::Int::new_const(&ctx, "y");
+    let y = ast::Int::new_const("y");
 
     let y_plus_one_greater_than_or_equal_to_two = y.clone().add(&one).ge(&two);
     goal.assert(&y_plus_one_greater_than_or_equal_to_two);
@@ -1080,12 +979,9 @@ fn check_application_of_tactic_to_goal() {
 
 #[test]
 fn test_goal_depth() {
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-
-    let goal = Goal::new(&ctx, false, false, false);
-    let a = ast::Bool::new_const(&ctx, "a");
-    let b = ast::Bool::new_const(&ctx, "b");
+    let goal = Goal::new(false, false, false);
+    let a = ast::Bool::new_const("a");
+    let b = ast::Bool::new_const("b");
     goal.assert(&a);
     goal.assert(&b);
     assert_eq!(goal.get_depth(), 0);
@@ -1093,12 +989,9 @@ fn test_goal_depth() {
 
 #[test]
 fn test_goal_size() {
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-
-    let goal = Goal::new(&ctx, false, false, false);
-    let a = ast::Bool::new_const(&ctx, "a");
-    let b = ast::Bool::new_const(&ctx, "b");
+    let goal = Goal::new(false, false, false);
+    let a = ast::Bool::new_const("a");
+    let b = ast::Bool::new_const("b");
     goal.assert(&a);
     goal.assert(&b);
     assert_eq!(goal.get_size(), 2);
@@ -1106,17 +999,14 @@ fn test_goal_size() {
 
 #[test]
 fn test_goal_num_expr() {
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-
-    let goal = Goal::new(&ctx, false, false, false);
-    let a = ast::Bool::new_const(&ctx, "a");
+    let goal = Goal::new(false, false, false);
+    let a = ast::Bool::new_const("a");
     goal.assert(&a);
     assert_eq!(goal.get_num_expr(), 1);
 
-    let goal = Goal::new(&ctx, false, false, false);
-    let a = ast::Bool::new_const(&ctx, "a");
-    let b = ast::Bool::new_const(&ctx, "b");
+    let goal = Goal::new(false, false, false);
+    let a = ast::Bool::new_const("a");
+    let b = ast::Bool::new_const("b");
     goal.assert(&a);
     goal.assert(&b);
     assert_eq!(goal.get_num_expr(), 2);
@@ -1124,44 +1014,36 @@ fn test_goal_num_expr() {
 
 #[test]
 fn test_goal_get_precision() {
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let false_bool = ast::Bool::from_bool(&ctx, false);
+    let false_bool = ast::Bool::from_bool(false);
 
-    let goal = Goal::new(&ctx, false, false, false);
+    let goal = Goal::new(false, false, false);
     goal.assert(&false_bool);
     assert_eq!(goal.get_precision(), z3::GoalPrec::Precise);
 }
 
 #[test]
 fn test_goal_is_inconsistent() {
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-
-    let false_bool = ast::Bool::from_bool(&ctx, false);
-    let goal = Goal::new(&ctx, false, false, false);
+    let false_bool = ast::Bool::from_bool(false);
+    let goal = Goal::new(false, false, false);
     goal.assert(&false_bool);
     assert!(goal.is_inconsistent());
 
-    let true_bool = ast::Bool::from_bool(&ctx, true);
-    let goal = Goal::new(&ctx, false, false, false);
+    let true_bool = ast::Bool::from_bool(true);
+    let goal = Goal::new(false, false, false);
     goal.assert(&true_bool);
     assert!(!goal.is_inconsistent());
 }
 
 #[test]
 fn test_goal_is_sat() {
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-
-    let false_bool = ast::Bool::from_bool(&ctx, false);
-    let goal = Goal::new(&ctx, false, false, false);
+    let false_bool = ast::Bool::from_bool(false);
+    let goal = Goal::new(false, false, false);
     goal.assert(&false_bool);
     assert!(!goal.is_decided_sat());
     assert!(goal.is_decided_unsat());
 
-    let true_bool = ast::Bool::from_bool(&ctx, true);
-    let goal = Goal::new(&ctx, false, false, false);
+    let true_bool = ast::Bool::from_bool(true);
+    let goal = Goal::new(false, false, false);
     goal.assert(&true_bool);
     assert!(!goal.is_decided_unsat());
     assert!(goal.is_decided_sat());
@@ -1169,11 +1051,8 @@ fn test_goal_is_sat() {
 
 #[test]
 fn test_goal_reset() {
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-
-    let a = ast::Bool::new_const(&ctx, "a");
-    let goal = Goal::new(&ctx, false, false, false);
+    let a = ast::Bool::new_const("a");
+    let goal = Goal::new(false, false, false);
     goal.assert(&a);
     assert_eq!(format!("{goal}"), "(goal\n  a)");
     goal.reset();
@@ -1183,17 +1062,16 @@ fn test_goal_reset() {
 #[test]
 fn test_set_membership() {
     let _ = env_logger::try_init();
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let solver = Solver::new(&ctx);
-    let set = ast::Set::new_const(&ctx, "integer_set", &Sort::int(&ctx));
-    let one = ast::Int::from_u64(&ctx, 1);
+
+    let solver = Solver::new();
+    let set = ast::Set::new_const("integer_set", &Sort::int());
+    let one = ast::Int::from_u64(1);
 
     solver.push();
-    solver.assert(&set.eq(&ast::Set::empty(&ctx, &Sort::int(&ctx))));
+    solver.assert(set._eq(ast::Set::empty(&Sort::int())));
 
     solver.push();
-    solver.assert(&set.member(&one));
+    solver.assert(set.member(&one));
     // An empty set will never contain 1
     assert_eq!(solver.check(), SatResult::Unsat);
     solver.pop(1);
@@ -1201,18 +1079,18 @@ fn test_set_membership() {
     solver.push();
     let set_with_one = set.add(&one);
     // Adding 1 to the empty set means it does now contain 1
-    solver.assert(&set_with_one.member(&one));
+    solver.assert(set_with_one.member(&one));
     assert_eq!(solver.check(), SatResult::Sat);
     let set_without_one = set_with_one.del(&one);
     // Removing 1 from the set means it no longer contains 1
-    solver.assert(&set_without_one.member(&one));
+    solver.assert(set_without_one.member(&one));
     assert_eq!(solver.check(), SatResult::Unsat);
     solver.pop(1);
 
     solver.push();
-    let x = ast::Int::new_const(&ctx, "x");
+    let x = ast::Int::new_const("x");
     // An empty set will always return false for member
-    let forall: ast::Bool = ast::forall_const(&ctx, &[&x], &[], &set.member(&x).not());
+    let forall: ast::Bool = ast::forall_const(&[&x], &[], &set.member(&x).not());
     solver.assert(&forall);
     assert_eq!(solver.check(), SatResult::Sat);
     solver.pop(1);
@@ -1221,8 +1099,8 @@ fn test_set_membership() {
 
     solver.push();
     // A singleton set of 1 will contain 1
-    solver.assert(&set.eq(&ast::Set::empty(&ctx, &Sort::int(&ctx)).add(&one)));
-    solver.assert(&set.member(&one));
+    solver.assert(set._eq(ast::Set::empty(&Sort::int()).add(&one)));
+    solver.assert(set.member(&one));
     assert_eq!(solver.check(), SatResult::Sat);
     solver.pop(1);
 }
@@ -1230,121 +1108,95 @@ fn test_set_membership() {
 #[test]
 fn test_dynamic_as_set() {
     let _ = env_logger::try_init();
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let set_sort = Sort::set(&ctx, &Sort::int(&ctx));
-    let array_sort = Sort::array(&ctx, &Sort::int(&ctx), &Sort::int(&ctx));
-    let array_of_sets = ast::Array::new_const(&ctx, "array_of_sets", &Sort::int(&ctx), &set_sort);
-    let array_of_arrays =
-        ast::Array::new_const(&ctx, "array_of_arrays", &Sort::int(&ctx), &array_sort);
-    assert!(array_of_sets
-        .select(&ast::Int::from_u64(&ctx, 0))
-        .as_set()
-        .is_some());
-    assert!(array_of_arrays
-        .select(&ast::Int::from_u64(&ctx, 0))
-        .as_set()
-        .is_none());
+
+    let set_sort = Sort::set(&Sort::int());
+    let array_sort = Sort::array(&Sort::int(), &Sort::int());
+    let array_of_sets = ast::Array::new_const("array_of_sets", &Sort::int(), &set_sort);
+    let array_of_arrays = ast::Array::new_const("array_of_arrays", &Sort::int(), &array_sort);
+    assert!(
+        array_of_sets
+            .select(&ast::Int::from_u64(0))
+            .as_set()
+            .is_some()
+    );
+    assert!(
+        array_of_arrays
+            .select(&ast::Int::from_u64(0))
+            .as_set()
+            .is_none()
+    );
 }
 
 #[test]
 fn test_array_store_select() {
     let _ = env_logger::try_init();
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let solver = Solver::new(&ctx);
-    let zero = ast::Int::from_u64(&ctx, 0);
-    let one = ast::Int::from_u64(&ctx, 1);
-    let set = ast::Array::new_const(&ctx, "integer_array", &Sort::int(&ctx), &Sort::int(&ctx))
-        .store(&zero, &one);
 
-    solver.assert(&set.select(&zero).eq(&one.into()).not());
+    let solver = Solver::new();
+    let zero = ast::Int::from_u64(0);
+    let one = ast::Int::from_u64(1);
+    let set = ast::Array::new_const("integer_array", &Sort::int(), &Sort::int()).store(&zero, &one);
+
+    solver.assert(set.select(&zero)._eq(one).not());
     assert_eq!(solver.check(), SatResult::Unsat);
 }
 
 #[test]
 fn test_goal_get_formulas() {
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-
-    let goal = Goal::new(&ctx, false, false, false);
-    let a = ast::Bool::new_const(&ctx, "a");
-    let b = ast::Bool::new_const(&ctx, "b");
-    let c = ast::Bool::new_const(&ctx, "c");
+    let goal = Goal::new(false, false, false);
+    let a = ast::Bool::new_const("a");
+    let b = ast::Bool::new_const("b");
+    let c = ast::Bool::new_const("c");
     goal.assert(&a);
     goal.assert(&b);
     goal.assert(&c);
-    let res = goal.get_formulas::<Bool>();
-    match res.as_slice() {
-        [ga, gb, gc] => {
-            assert!(ga.ast_eq(&a));
-            assert!(gb.ast_eq(&b));
-            assert!(gc.ast_eq(&c));
-        }
-        x => panic!("{:?}", x),
-    }
+    assert_eq!(goal.get_formulas::<Bool>(), vec![a, b, c],);
 }
 
 #[test]
 fn test_tactic_skip() {
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let params = Params::new(&ctx);
+    let params = Params::new();
 
-    let a = ast::Bool::new_const(&ctx, "a");
-    let b = ast::Bool::new_const(&ctx, "b");
-    let bools = [&a, &b, &a];
-    let a_and_b_and_a = Bool::and(&ctx, &bools);
-    let goal = Goal::new(&ctx, false, false, false);
+    let a = ast::Bool::new_const("a");
+    let b = ast::Bool::new_const("b");
+    let a_and_b_and_a = &a & &b & &a;
+    let goal = Goal::new(false, false, false);
     goal.assert(&a_and_b_and_a);
 
-    let tactic = Tactic::create_skip(&ctx);
+    let tactic = Tactic::create_skip();
     let apply_results = tactic.apply(&goal, Some(&params));
     let goal_results = apply_results
         .unwrap()
         .list_subgoals()
         .collect::<Vec<Goal>>();
     let goal_result = goal_results.first().unwrap();
-    let res = goal_result.get_formulas::<Bool>();
-    match res.as_slice() {
-        [ga, gb, gc] => {
-            assert!(ga.ast_eq(&a.clone()));
-            assert!(gb.ast_eq(&b));
-            assert!(gc.ast_eq(&a));
-        }
-        x => panic!("{:?}", x),
-    }
+    assert_eq!(goal_result.get_formulas::<Bool>(), vec![a.clone(), b, a],);
 }
 
 #[test]
 fn test_tactic_fail() {
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let params = Params::new(&ctx);
+    let params = Params::new();
 
-    let a = ast::Bool::new_const(&ctx, "a");
-    let goal = Goal::new(&ctx, false, false, false);
+    let a = ast::Bool::new_const("a");
+    let goal = Goal::new(false, false, false);
     goal.assert(&a);
 
-    let tactic = Tactic::new(&ctx, "fail");
+    let tactic = Tactic::new("fail");
     let apply_results = tactic.apply(&goal, Some(&params));
     assert!(apply_results.is_err());
 }
 
 #[test]
 fn test_tactic_try_for() {
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let params = Params::new(&ctx);
+    let params = Params::new();
 
-    let one = ast::Int::from_i64(&ctx, 1);
-    let two = ast::Int::from_i64(&ctx, 2);
-    let x = ast::Int::new_const(&ctx, "x");
+    let one = ast::Int::from_i64(1);
+    let two = ast::Int::from_i64(2);
+    let x = ast::Int::new_const("x");
 
-    let goal = Goal::new(&ctx, false, false, false);
-    goal.assert(&x.ge(&one.add(&two)));
+    let goal = Goal::new(false, false, false);
+    goal.assert(&x.ge(one.add(&two)));
 
-    let tactic = Tactic::new(&ctx, "simplify");
+    let tactic = Tactic::new("simplify");
     let try_for_tactic = tactic.try_for(Duration::from_secs(u64::MAX));
 
     // Test that `try_for` can successfully apply the underlying tactic
@@ -1359,43 +1211,37 @@ fn test_tactic_try_for() {
 
 #[test]
 fn test_tactic_and_then() {
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let params = Params::new(&ctx);
+    let params = Params::new();
 
-    let a = ast::Bool::new_const(&ctx, "a");
-    let b = ast::Bool::new_const(&ctx, "b");
-    let bools = [&a, &b, &a];
-    let a_and_b_and_a = Bool::and(&ctx, &bools);
-    let goal = Goal::new(&ctx, false, false, false);
+    let a = ast::Bool::new_const("a");
+    let b = ast::Bool::new_const("b");
+    let a_and_b_and_a = &a & &b & &a;
+    let goal = Goal::new(false, false, false);
     goal.assert(&a_and_b_and_a);
 
-    let tactic = Tactic::new(&ctx, "sat-preprocess");
-    let and_then_tactic = tactic.and_then(&Tactic::new(&ctx, "simplify"));
+    let tactic = Tactic::new("sat-preprocess");
+    let and_then_tactic = tactic.and_then(&Tactic::new("simplify"));
     let apply_results = and_then_tactic.apply(&goal, Some(&params));
     let goal_results = apply_results
         .unwrap()
         .list_subgoals()
         .collect::<Vec<Goal>>();
     let goal_result = goal_results.first().unwrap();
-    assert!(slice_ast_eq(&goal_result.get_formulas::<Bool>(), &[a, b]));
+    assert_eq!(goal_result.get_formulas::<Bool>(), vec![a, b]);
 }
 
 #[test]
 fn test_tactic_or_else() {
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let params = Params::new(&ctx);
+    let params = Params::new();
 
-    let a = ast::Bool::new_const(&ctx, "a");
-    let b = ast::Bool::new_const(&ctx, "b");
-    let bools = [&a, &b, &a];
-    let a_and_b_and_a = Bool::and(&ctx, &bools);
-    let goal = Goal::new(&ctx, false, false, false);
+    let a = ast::Bool::new_const("a");
+    let b = ast::Bool::new_const("b");
+    let a_and_b_and_a = &a & &b & &a;
+    let goal = Goal::new(false, false, false);
     goal.assert(&a_and_b_and_a);
 
-    let tactic = Tactic::new(&ctx, "sat-preprocess");
-    let simplify = Tactic::new(&ctx, "simplify");
+    let tactic = Tactic::new("sat-preprocess");
+    let simplify = Tactic::new("simplify");
     let or_else_tactic = tactic.or_else(&simplify);
     let apply_results = or_else_tactic.apply(&goal, Some(&params));
     let goal_results = apply_results
@@ -1403,275 +1249,207 @@ fn test_tactic_or_else() {
         .list_subgoals()
         .collect::<Vec<Goal>>();
     let goal_result = goal_results.first().unwrap();
-    assert!(slice_ast_eq(&goal_result.get_formulas::<Bool>(), &[a, b]));
+    assert_eq!(goal_result.get_formulas::<Bool>(), vec![a, b]);
 }
 
 #[test]
 fn test_goal_apply_tactic() {
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
+    pub fn test_apply_tactic(goal: Goal, before_formulas: Vec<Bool>, after_formulas: Vec<Bool>) {
+        assert_eq!(goal.get_formulas::<Bool>(), before_formulas);
+        let params = Params::new();
 
-    pub fn test_apply_tactic(
-        ctx: &Context,
-        goal: Goal,
-        before_formulas: Vec<Bool>,
-        after_formulas: Vec<Bool>,
-    ) {
-        assert!(slice_ast_eq(&goal.get_formulas::<Bool>(), &before_formulas));
-        let params = Params::new(ctx);
-
-        let tactic = Tactic::new(ctx, "sat-preprocess");
-        let repeat_tactic = Tactic::repeat(ctx, &tactic, 100);
+        let tactic = Tactic::new("sat-preprocess");
+        let repeat_tactic = Tactic::repeat(&tactic, 100);
         let apply_results = repeat_tactic.apply(&goal, Some(&params));
         let goal_results = apply_results
             .unwrap()
             .list_subgoals()
             .collect::<Vec<Goal>>();
         let goal_result = goal_results.first().unwrap();
-        assert!(slice_ast_eq(
-            &goal_result.get_formulas::<Bool>(),
-            &after_formulas
-        ));
+        assert_eq!(
+            goal_result.get_formulas::<Bool>(),
+            after_formulas,
+            "Before: {before_formulas:?}"
+        );
     }
 
-    let a = ast::Bool::new_const(&ctx, "a");
-    let b = ast::Bool::new_const(&ctx, "b");
+    let a = ast::Bool::new_const("a");
+    let b = ast::Bool::new_const("b");
 
-    let bools = [&a, &b, &a];
-    let a_and_b_and_a = Bool::and(&ctx, &bools);
-    let goal = Goal::new(&ctx, false, false, false);
+    let a_and_b_and_a = &a & &b & &a;
+    let goal = Goal::new(false, false, false);
     goal.assert(&a_and_b_and_a);
     test_apply_tactic(
-        &ctx,
         goal,
         vec![a.clone(), b.clone(), a.clone()],
         vec![a.clone(), b.clone()],
     );
 
     let a_implies_b = ast::Bool::implies(&a, &b).simplify();
-    let a_and_a_implies_b = Bool::and(&ctx, &[&a, &a_implies_b]);
+    let a_and_a_implies_b = Bool::and(&[&a, &a_implies_b]);
 
-    let goal = Goal::new(&ctx, false, false, false);
+    let goal = Goal::new(false, false, false);
     goal.assert(&a_and_a_implies_b);
     test_apply_tactic(
-        &ctx,
         goal,
         vec![a.clone(), a_implies_b.clone()],
         vec![a.clone(), b.clone()],
     );
 
-    let goal = Goal::new(&ctx, false, false, false);
+    let goal = Goal::new(false, false, false);
     goal.assert(&a);
     goal.assert(&a_implies_b.clone());
     test_apply_tactic(
-        &ctx,
         goal,
         vec![a.clone(), a_implies_b.clone()],
         vec![a.clone(), b.clone()],
     );
 
-    let true_bool = ast::Bool::from_bool(&ctx, true);
-    let false_bool = ast::Bool::from_bool(&ctx, false);
-    let goal = Goal::new(&ctx, false, false, false);
-    let true_and_false_and_true = ast::Bool::and(&ctx, &[&true_bool, &false_bool, &true_bool]);
+    let false_bool = ast::Bool::from_bool(false);
+    let goal = Goal::new(false, false, false);
+    let true_and_false_and_true = ast::Bool::and(&[true, false, true]);
     goal.assert(&true_and_false_and_true);
-    test_apply_tactic(
-        &ctx,
-        goal,
-        vec![false_bool.clone()],
-        vec![false_bool.clone()],
-    );
+    test_apply_tactic(goal, vec![false_bool.clone()], vec![false_bool.clone()]);
 }
 
 #[test]
 fn test_tactic_cond() {
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let t1 = Tactic::new(&ctx, "qfnra");
-    let t2 = Tactic::new(&ctx, "smt");
-    let p = Probe::new(&ctx, "is-qfnra");
+    let t1 = Tactic::new("qfnra");
+    let t2 = Tactic::new("smt");
+    let p = Probe::new("is-qfnra");
 
-    let _t = Tactic::cond(&ctx, &p, &t1, &t2);
+    let _t = Tactic::cond(&p, &t1, &t2);
 }
 
 #[test]
 fn test_tactic_conditions() {
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let t1 = Tactic::new(&ctx, "qfnra");
-    let t2 = Tactic::new(&ctx, "smt");
-    let p = Probe::new(&ctx, "is-qfnra");
+    let t1 = Tactic::new("qfnra");
+    let t2 = Tactic::new("smt");
+    let p = Probe::new("is-qfnra");
 
     t1.probe_or_else(&p, &t2);
     t1.when(&p);
-    Tactic::cond(&ctx, &p, &t1, &t2);
-    Tactic::fail_if(&ctx, &p);
+    Tactic::cond(&p, &t1, &t2);
+    Tactic::fail_if(&p);
 }
 
 #[test]
 fn test_probe_debug() {
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let _v: Vec<&str> = Probe::list_all(&ctx).map(|x| x.unwrap()).collect();
+    let _v: Vec<String> = Probe::list_all().into_iter().map(|x| x.unwrap()).collect();
     assert_eq!(
         "A probe to give an upper bound of Ackermann congruence lemmas that a formula might generate.",
-        Probe::describe(&ctx, "ackr-bound-probe").unwrap(),
+        Probe::describe("ackr-bound-probe").unwrap(),
     );
 }
 
 #[test]
 fn test_probe_names() {
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
+    let x = ast::Int::fresh_const("x");
+    let g = Goal::new(false, false, false);
 
-    let x = ast::Int::fresh_const(&ctx, "x");
-    let zero = ast::Int::from_u64(&ctx, 0);
-    let ten = ast::Int::from_u64(&ctx, 10);
-    let twenty = ast::Int::from_u64(&ctx, 20);
-    let g = Goal::new(&ctx, false, false, false);
+    g.assert(&x.gt(0));
+    g.assert(&x.lt(10));
+    g.assert(&x.lt(20));
 
-    g.assert(&x.gt(&zero));
-    g.assert(&x.lt(&ten));
-    g.assert(&x.lt(&twenty));
-
-    let num_consts_probe = Probe::new(&ctx, "num-consts");
+    let num_consts_probe = Probe::new("num-consts");
     assert_eq!(1.0, num_consts_probe.apply(&g));
 
-    let is_prop_probe = Probe::new(&ctx, "is-propositional");
+    let is_prop_probe = Probe::new("is-propositional");
     assert_eq!(0.0, is_prop_probe.apply(&g));
 
-    let is_qflia_probe = Probe::new(&ctx, "is-qflia");
+    let is_qflia_probe = Probe::new("is-qflia");
     assert_eq!(1.0, is_qflia_probe.apply(&g));
 }
 
 #[test]
 fn test_probe_eq() {
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-
-    let zero = ast::Int::from_u64(&ctx, 0);
-    let ten = ast::Int::from_u64(&ctx, 10);
-
-    let two_probe = Probe::constant(&ctx, 2.0);
-    let size_probe = Probe::new(&ctx, "size");
+    let two_probe = Probe::constant(2.0);
+    let size_probe = Probe::new("size");
     let equals_two_probe = &size_probe.eq(&two_probe);
 
-    let x = ast::Int::fresh_const(&ctx, "x");
-    let g = Goal::new(&ctx, false, false, false);
-    g.assert(&x.gt(&zero));
-    g.assert(&x.lt(&ten));
+    let x = ast::Int::fresh_const("x");
+    let g = Goal::new(false, false, false);
+    g.assert(&x.gt(0));
+    g.assert(&x.lt(10));
     assert_eq!(1.0, equals_two_probe.apply(&g));
 }
 
 #[test]
 fn test_probe_gt() {
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let ten_probe = Probe::constant(&ctx, 10.0);
-    let size_probe = Probe::new(&ctx, "size");
+    let ten_probe = Probe::constant(10.0);
+    let size_probe = Probe::new("size");
     let gt_ten_probe = &size_probe.gt(&ten_probe);
 
-    let x = ast::Int::fresh_const(&ctx, "x");
-    let zero = ast::Int::from_u64(&ctx, 0);
-    let ten = ast::Int::from_u64(&ctx, 10);
-    let g = Goal::new(&ctx, false, false, false);
-    g.assert(&x.gt(&zero));
-    g.assert(&x.lt(&ten));
+    let x = ast::Int::fresh_const("x");
+    let g = Goal::new(false, false, false);
+    g.assert(&x.gt(0));
+    g.assert(&x.lt(10));
     assert_eq!(0.0, gt_ten_probe.apply(&g));
 }
 
 #[test]
 fn test_probe_gte() {
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let two_probe = Probe::constant(&ctx, 2.0);
-    let size_probe = Probe::new(&ctx, "size");
+    let two_probe = Probe::constant(2.0);
+    let size_probe = Probe::new("size");
     let ge_two_probe = &size_probe.ge(&two_probe);
 
-    let x = ast::Int::fresh_const(&ctx, "x");
-    let zero = ast::Int::from_u64(&ctx, 0);
-    let ten = ast::Int::from_u64(&ctx, 10);
-    let g = Goal::new(&ctx, false, false, false);
-    g.assert(&x.gt(&zero));
-    g.assert(&x.lt(&ten));
+    let x = ast::Int::fresh_const("x");
+    let g = Goal::new(false, false, false);
+    g.assert(&x.gt(0));
+    g.assert(&x.lt(10));
     assert_eq!(1.0, ge_two_probe.apply(&g));
 }
 
 #[test]
 fn test_probe_le() {
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let two_probe = Probe::constant(&ctx, 2.0);
-    let size_probe = Probe::new(&ctx, "size");
+    let two_probe = Probe::constant(2.0);
+    let size_probe = Probe::new("size");
     let le_two_probe = &size_probe.le(&two_probe);
 
-    let x = ast::Int::fresh_const(&ctx, "x");
-    let zero = ast::Int::from_u64(&ctx, 0);
-    let ten = ast::Int::from_u64(&ctx, 10);
-    let g = Goal::new(&ctx, false, false, false);
-    g.assert(&x.gt(&zero));
-    g.assert(&x.lt(&ten));
+    let x = ast::Int::fresh_const("x");
+    let g = Goal::new(false, false, false);
+    g.assert(&x.gt(0));
+    g.assert(&x.lt(10));
     assert_eq!(1.0, le_two_probe.apply(&g));
 }
 
 #[test]
 fn test_probe_lt() {
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let ten_probe = Probe::constant(&ctx, 10.0);
-    let size_probe = Probe::new(&ctx, "size");
+    let ten_probe = Probe::constant(10.0);
+    let size_probe = Probe::new("size");
     let le_ten_probe = &size_probe.le(&ten_probe);
 
-    let x = ast::Int::fresh_const(&ctx, "x");
-    let zero = ast::Int::from_u64(&ctx, 0);
-    let ten = ast::Int::from_u64(&ctx, 10);
-    let g = Goal::new(&ctx, false, false, false);
-    g.assert(&x.gt(&zero));
-    g.assert(&x.lt(&ten));
+    let x = ast::Int::fresh_const("x");
+    let g = Goal::new(false, false, false);
+    g.assert(&x.gt(0));
+    g.assert(&x.lt(10));
     assert_eq!(1.0, le_ten_probe.apply(&g));
 }
 
 #[test]
 fn test_probe_ne() {
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let two_probe = Probe::constant(&ctx, 2.0);
-    let size_probe = Probe::new(&ctx, "size");
+    let two_probe = Probe::constant(2.0);
+    let size_probe = Probe::new("size");
     let ne_two_probe = &size_probe.ne(&two_probe);
 
-    let x = ast::Int::fresh_const(&ctx, "x");
-    let zero = ast::Int::from_u64(&ctx, 0);
-    let ten = ast::Int::from_u64(&ctx, 10);
-    let g = Goal::new(&ctx, false, false, false);
-    g.assert(&x.gt(&zero));
-    g.assert(&x.lt(&ten));
+    let x = ast::Int::fresh_const("x");
+    let g = Goal::new(false, false, false);
+    g.assert(&x.gt(0));
+    g.assert(&x.lt(10));
     assert_eq!(0.0, ne_two_probe.apply(&g));
 }
 
 #[test]
-#[should_panic]
-fn test_issue_94() {
-    let cfg = Config::new();
-    let ctx0 = Context::new(&cfg);
-    let ctx1 = Context::new(&cfg);
-    let i0 = ast::Int::fresh_const(&ctx0, "a");
-    let i1 = ast::Int::fresh_const(&ctx1, "b");
-    ast::Int::add(&ctx0, &[&i0, &i1]);
-}
-
-#[test]
 fn test_ast_safe_eq() {
-    let cfg = Config::new();
-    let ctx = &Context::new(&cfg);
-    let x: ast::Dynamic = ast::Bool::new_const(ctx, "a").into();
-    let y: ast::Dynamic = ast::String::from_str(ctx, "b").unwrap().into();
+    let x: ast::Dynamic = ast::Bool::new_const("a").into();
+    let y: ast::Dynamic = ast::String::from_str("b").unwrap().into();
 
-    let other_bool: ast::Dynamic = ast::Bool::new_const(ctx, "c").into();
-    let other_string: ast::Dynamic = ast::String::from_str(ctx, "d").unwrap().into();
+    let other_bool: ast::Dynamic = ast::Bool::new_const("c").into();
+    let other_string: ast::Dynamic = ast::String::from_str("d").unwrap().into();
 
-    let sd: SortDiffers<'_> = SortDiffers::new(other_bool.get_sort(), other_string.get_sort());
+    let sd: SortDiffers = SortDiffers::new(other_bool.get_sort(), other_string.get_sort());
 
-    let result = x.safe_eq(&y);
+    let result = x._safe_eq(&y);
     assert!(result.is_err());
     let err = result.err().unwrap();
     assert_eq!(err.left(), sd.left());
@@ -1680,17 +1458,15 @@ fn test_ast_safe_eq() {
 
 #[test]
 fn test_ast_safe_decl() {
-    let cfg = Config::new();
-    let ctx = &Context::new(&cfg);
-    let x: ast::Bool = ast::Bool::new_const(ctx, "x");
+    let x: ast::Bool = ast::Bool::new_const("x");
     let x_not = x.not();
     assert_eq!(x_not.safe_decl().unwrap().kind(), DeclKind::NOT);
 
-    let f = FuncDecl::new(ctx, "f", &[&Sort::int(ctx)], &Sort::int(ctx));
-    let x = ast::Int::new_const(ctx, "x");
+    let f = FuncDecl::new("f", &[&Sort::int()], &Sort::int());
+    let x = ast::Int::new_const("x");
     let f_x: ast::Int = f.apply(&[&x]).try_into().unwrap();
-    let f_x_pattern: Pattern = Pattern::new(ctx, &[&f_x]);
-    let forall = ast::forall_const(ctx, &[&x], &[&f_x_pattern], &x.eq(&f_x));
+    let f_x_pattern: Pattern = Pattern::new(&[&f_x]);
+    let forall = ast::forall_const(&[&x], &[&f_x_pattern], &x._eq(&f_x));
     assert!(forall.safe_decl().is_err());
     assert_eq!(
         format!("{}", forall.safe_decl().err().unwrap()),
@@ -1701,89 +1477,59 @@ fn test_ast_safe_decl() {
 //the intersection of "FOO"+"bar" and [a-z]+ is empty
 #[test]
 fn test_regex_capital_foobar_intersect_az_plus_is_unsat() {
-    let cfg = Config::new();
-    let ctx = &Context::new(&cfg);
-    let solver = Solver::new(ctx);
-    let s = ast::String::new_const(ctx, "s");
+    let solver = Solver::new();
+    let s = ast::String::new_const("s");
 
-    let re = ast::Regexp::intersect(
-        ctx,
-        &[
-            &ast::Regexp::concat(
-                ctx,
-                &[
-                    &ast::Regexp::literal(ctx, "FOO"),
-                    &ast::Regexp::literal(ctx, "bar"),
-                ],
-            ),
-            &ast::Regexp::plus(&ast::Regexp::range(ctx, &'a', &'z')),
-        ],
-    );
-    solver.assert(&s.regex_matches(&re));
+    let re = ast::Regexp::intersect(&[
+        &ast::Regexp::concat(&[&ast::Regexp::literal("FOO"), &ast::Regexp::literal("bar")]),
+        &ast::Regexp::plus(&ast::Regexp::range(&'a', &'z')),
+    ]);
+    solver.assert(s.regex_matches(&re));
     assert!(solver.check() == SatResult::Unsat);
 }
 
 #[test]
 fn test_regex_union() {
-    let cfg = Config::new();
-    let ctx = &Context::new(&cfg);
-    let solver = Solver::new(ctx);
-    let a = ast::String::from_str(ctx, "a").unwrap();
-    let b = ast::String::from_str(ctx, "b").unwrap();
-    let c = ast::String::from_str(ctx, "c").unwrap();
-    let re = ast::Regexp::union(
-        ctx,
-        &[
-            ast::Regexp::literal(ctx, "a"),
-            ast::Regexp::literal(ctx, "b"),
-        ],
-    );
-    solver.assert(&a.regex_matches(&re));
-    solver.assert(&b.regex_matches(&re));
-    solver.assert(&c.regex_matches(&re).not());
+    let solver = Solver::new();
+    let a = ast::String::from_str("a").unwrap();
+    let b = ast::String::from_str("b").unwrap();
+    let c = ast::String::from_str("c").unwrap();
+    let re = ast::Regexp::union(&[ast::Regexp::literal("a"), ast::Regexp::literal("b")]);
+    solver.assert(a.regex_matches(&re));
+    solver.assert(b.regex_matches(&re));
+    solver.assert(c.regex_matches(&re).not());
     assert!(solver.check() == SatResult::Sat);
 }
 
 #[test]
 fn test_regex_union2() {
-    let cfg = Config::new();
-    let ctx = &Context::new(&cfg);
-    let solver = Solver::new(ctx);
-    let c = ast::String::from_str(ctx, "c").unwrap();
-    let re = ast::Regexp::union(
-        ctx,
-        &[
-            ast::Regexp::literal(ctx, "a"),
-            ast::Regexp::literal(ctx, "b"),
-        ],
-    );
-    solver.assert(&c.regex_matches(&re));
+    let solver = Solver::new();
+    let c = ast::String::from_str("c").unwrap();
+    let re = ast::Regexp::union(&[ast::Regexp::literal("a"), ast::Regexp::literal("b")]);
+    solver.assert(c.regex_matches(&re));
     assert!(solver.check() == SatResult::Unsat);
 }
 
 #[test]
 /// <https://github.com/Z3Prover/z3/blob/21e59f7c6e5033006265fc6bc16e2c9f023db0e8/examples/dotnet/Program.cs#L329-L370>
 fn test_array_example1() {
-    let cfg = Config::new();
-    let ctx = &Context::new(&cfg);
+    let g = Goal::new(true, false, false);
 
-    let g = Goal::new(ctx, true, false, false);
+    let aex = Array::new_const("MyArray", &Sort::int(), &Sort::bitvector(32));
 
-    let aex = Array::new_const(ctx, "MyArray", &Sort::int(ctx), &Sort::bitvector(ctx, 32));
+    let sel = aex.select(&Int::from_u64(0));
 
-    let sel = aex.select(&Int::from_u64(ctx, 0));
+    g.assert(&sel._eq(BV::from_u64(42, 32)));
 
-    g.assert(&sel.eq(&BV::from_u64(ctx, 42, 32).into()));
+    let xc = Int::new_const("x");
 
-    let xc = Int::new_const(ctx, "x");
-
-    let fd = FuncDecl::new(ctx, "f", &[&Sort::int(ctx)], &Sort::int(ctx));
+    let fd = FuncDecl::new("f", &[&Sort::int()], &Sort::int());
 
     let fapp = fd.apply(&[&xc as &dyn Ast]);
 
-    g.assert(&Int::from_u64(ctx, 123).eq(&xc.clone().add(&fapp.as_int().unwrap())));
+    g.assert(&Int::from_u64(123)._eq(xc.clone().add(&fapp.as_int().unwrap())));
 
-    let s = &Solver::new(ctx);
+    let s = &Solver::new();
     for a in g.get_formulas() {
         s.assert(&a);
     }
@@ -1816,27 +1562,20 @@ fn test_array_example1() {
 #[test]
 /// <https://z3prover.github.io/api/html/classz3py_1_1_func_entry.html>
 fn return_number_args_in_given_entry() {
-    let cfg = Config::new();
-    let ctx = &Context::new(&cfg);
-    let f = FuncDecl::new(
-        ctx,
-        "f",
-        &[&Sort::int(ctx), &Sort::int(ctx)],
-        &Sort::int(ctx),
-    );
+    let f = FuncDecl::new("f", &[&Sort::int(), &Sort::int()], &Sort::int());
 
-    let solver = Solver::new(ctx);
+    let solver = Solver::new();
     solver.assert(
-        &f.apply(&[&Int::from_u64(ctx, 0), &Int::from_u64(ctx, 1)])
-            .eq(&Int::from_u64(ctx, 10).into()),
+        f.apply(&[&Int::from_u64(0), &Int::from_u64(1)])
+            ._eq(Int::from_u64(10)),
     );
     solver.assert(
-        &f.apply(&[&Int::from_u64(ctx, 1), &Int::from_u64(ctx, 2)])
-            .eq(&Int::from_u64(ctx, 20).into()),
+        f.apply(&[&Int::from_u64(1), &Int::from_u64(2)])
+            ._eq(Int::from_u64(20)),
     );
     solver.assert(
-        &f.apply(&[&Int::from_u64(ctx, 1), &Int::from_u64(ctx, 0)])
-            .eq(&Int::from_u64(ctx, 10).into()),
+        f.apply(&[&Int::from_u64(1), &Int::from_u64(0)])
+            ._eq(Int::from_u64(10)),
     );
 
     assert!(solver.check() == SatResult::Sat);
@@ -1857,20 +1596,16 @@ fn return_number_args_in_given_entry() {
 #[test]
 /// <https://stackoverflow.com/questions/13395391/z3-finding-all-satisfying-models>
 fn iterate_all_solutions() {
-    let cfg = Config::new();
-    let ctx = &Context::new(&cfg);
-    let solver = Solver::new(ctx);
-    let a = &Int::new_const(ctx, "a");
-    let b = &Int::new_const(ctx, "b");
-    let one = Int::from_u64(ctx, 1);
-    let two = Int::from_u64(ctx, 2);
-    let five = &Int::from_u64(ctx, 5);
+    let solver = Solver::new();
+    let a = &Int::new_const("a");
+    let b = &Int::new_const("b");
+    let one = Int::from_u64(1);
 
-    solver.assert(&one.le(a));
-    solver.assert(&a.le(five));
-    solver.assert(&one.le(b));
-    solver.assert(&b.le(five));
-    solver.assert(&a.ge(&(two * b)));
+    solver.assert(one.le(a));
+    solver.assert(a.le(5));
+    solver.assert(one.le(b));
+    solver.assert(b.le(5));
+    solver.assert(a.ge(&(2 * b)));
 
     let mut solutions = std::collections::HashSet::new();
     while solver.check() == SatResult::Sat {
@@ -1881,7 +1616,7 @@ fn iterate_all_solutions() {
             .map(|fd| {
                 modifications.push(
                     fd.apply(&[])
-                        .eq(&model.get_const_interp(&fd.apply(&[])).unwrap())
+                        ._eq(model.get_const_interp(&fd.apply(&[])).unwrap())
                         .not(),
                 );
                 format!(
@@ -1893,7 +1628,7 @@ fn iterate_all_solutions() {
             .collect::<Vec<_>>()
             .join(", ");
         solutions.insert(format!("[{this_solution}]"));
-        solver.assert(&Bool::or(ctx, &modifications.iter().collect::<Vec<_>>()));
+        solver.assert(Bool::or(&modifications.iter().collect::<Vec<_>>()));
     }
 
     assert!(
@@ -1915,4 +1650,139 @@ fn iterate_all_solutions() {
 fn get_version() {
     println!("Z3 version: {:?}", z3::version());
     println!("Z3 full version string: {}", z3::full_version());
+}
+
+#[test]
+fn test_consequences() {
+    let solver = Solver::new();
+    let a = Bool::new_const("a");
+    let b = Bool::new_const("b");
+    let c = Bool::new_const("c");
+    let d = Bool::new_const("d");
+    solver.assert(a.implies(&b));
+    solver.assert(b.implies(&c));
+
+    let assumptions = vec![a.clone()];
+    let variables = vec![b.clone(), c.clone(), d.clone()];
+    let mut cons = solver.get_consequences(&assumptions, &variables);
+    assert!(cons.len() == 2);
+    assert!(cons.pop().unwrap().to_string() == "(=> a c)");
+    assert!(cons.pop().unwrap().to_string() == "(=> a b)");
+
+    let assumptions = vec![c.not(), d.clone()];
+    let variables = vec![a, b, c, d];
+    let mut cons = solver.get_consequences(&assumptions, &variables);
+    assert!(cons.len() == 4);
+    assert!(cons.pop().unwrap().to_string() == "(=> (not c) (not a))");
+    assert!(cons.pop().unwrap().to_string() == "(=> (not c) (not b))");
+    assert!(cons.pop().unwrap().to_string() == "(=> (not c) (not c))");
+    assert!(cons.pop().unwrap().to_string() == "(=> d d)");
+}
+
+#[test]
+fn test_atmost() {
+    let solver = Solver::new();
+    let a = Bool::new_const("a");
+    let b = Bool::new_const("b");
+    let c = Bool::new_const("c");
+    let d = Bool::new_const("d");
+    solver.assert(a.implies(&b));
+    solver.assert(b.implies(&c));
+
+    solver.push();
+    let am = atmost([&a, &b, &c, &d], 2);
+    solver.assert(&am);
+    assert!(matches!(solver.check(), SatResult::Sat));
+    solver.pop(1);
+
+    solver.push();
+    solver.assert(&a);
+    let am = atmost([&a, &b, &c, &d], 0);
+    solver.assert(&am);
+    assert!(matches!(solver.check(), SatResult::Unsat));
+    solver.pop(1);
+}
+
+#[test]
+fn test_atleast() {
+    let solver = Solver::new();
+    let a = Bool::new_const("a");
+    let b = Bool::new_const("b");
+    let c = Bool::new_const("c");
+    let d = Bool::new_const("d");
+    solver.assert(a.implies(&b));
+    solver.assert(b.implies(&c));
+
+    solver.push();
+    let am = atleast([&a, &b, &c, &d], 4);
+    solver.assert(&am);
+    assert!(matches!(solver.check(), SatResult::Sat));
+    solver.pop(1);
+
+    solver.push();
+    solver.assert(a.not());
+    let am = atleast([&a, &b, &c, &d], 4);
+    solver.assert(&am);
+    assert!(matches!(solver.check(), SatResult::Unsat));
+    solver.pop(1);
+}
+
+#[test]
+fn test_model_iter() {
+    let solver = Solver::new();
+
+    let a = ast::Int::new_const("a");
+    solver.assert(a._eq(2));
+    solver.check();
+
+    let model = solver.get_model().unwrap();
+
+    // consume a model and return some asts constructed from it
+    // this didn't compile before pr#324
+    fn consume_model(model: Model) -> Vec<ast::Dynamic> {
+        let mut asts = vec![];
+        for func in &model {
+            asts.push(func.apply(&[]));
+        }
+        asts
+    }
+
+    assert_eq!(
+        consume_model(model),
+        vec![ast::Dynamic::new_const("a", &Sort::int())]
+    );
+}
+
+#[test]
+fn test_round_towards_nearest_away() {
+    let solver = Solver::new();
+
+    let eps = f64::from_bits(0x3cb0000000000000);
+    let x = z3::ast::Float::from_f64(1.0);
+    let y = z3::ast::Float::from_f64(0.5 * eps);
+
+    let rtna = z3::ast::RoundingMode::round_nearest_ties_to_away();
+    let res_rtna = x.add_with_rounding_mode(y, &rtna);
+
+    let expected = z3::ast::Float::from_f64(1.0000000000000002);
+    solver.assert(res_rtna._eq(&expected));
+
+    assert_eq!(solver.check(), SatResult::Sat);
+}
+
+#[test]
+fn test_round_towards_nearest_even() {
+    let solver = Solver::new();
+
+    let eps = f64::from_bits(0x3cb0000000000000);
+    let x = z3::ast::Float::from_f64(1.0);
+    let y = z3::ast::Float::from_f64(0.5 * eps);
+
+    let rtne = z3::ast::RoundingMode::round_nearest_ties_to_even();
+    let res_rtne = x.add_with_rounding_mode(y, &rtne);
+
+    let expected = z3::ast::Float::from_f64(1.0);
+    solver.assert(res_rtne._eq(&expected));
+
+    assert_eq!(solver.check(), SatResult::Sat);
 }
