@@ -140,9 +140,7 @@ fn test_solve_simple_semver_example() {
         .collect(),
     );
 
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let opt = Optimize::new(&ctx);
+    let opt = Optimize::new();
 
     let mut root: HashMap<String, VersionReq> = HashMap::new();
     let mut asts: HashMap<String, ast::Int> = HashMap::new();
@@ -156,21 +154,21 @@ fn test_solve_simple_semver_example() {
 
     // Make a root Z3 Int constant for each pkg we're trying to solve for.
     for (k, v) in &root {
-        let ast = ast::Int::fresh_const(&ctx, "root-pkg");
-        info!("new AST for root {}", k);
+        let ast = ast::Int::fresh_const("root-pkg");
+        info!("new AST for root {k}");
 
         match first_version_req_index(&smap, k, v) {
             None => (),
             Some(low) => {
-                info!("Asserting: {} >= #{} (root)", k, low);
-                opt.assert(&ast.ge(&ast::Int::from_u64(&ctx, low as u64)));
+                info!("Asserting: {k} >= #{low} (root)");
+                opt.assert(&ast.ge(ast::Int::from_u64(low as u64)));
             }
         }
         match last_version_req_index(&smap, k, v) {
             None => (),
             Some(high) => {
-                info!("Asserting: {} <= #{} (root)", k, high);
-                opt.assert(&ast.le(&ast::Int::from_u64(&ctx, high as u64)));
+                info!("Asserting: {k} <= #{high} (root)");
+                opt.assert(&ast.le(ast::Int::from_u64(high as u64)));
             }
         }
         asts.insert(k.clone(), ast);
@@ -178,23 +176,22 @@ fn test_solve_simple_semver_example() {
 
     // Tell the optimizer to maximizes the sum of the root constants.
     opt.maximize(&ast::Int::add(
-        &ctx,
         &asts.values().collect::<Vec<&ast::Int>>()[..],
     ));
 
     // Ensure we have a constant for every pkg _or_ dep listed
     for k in (smap).keys() {
         asts.entry(k.clone()).or_insert_with(|| {
-            info!("new AST for {}", k);
-            ast::Int::fresh_const(&ctx, "pkg")
+            info!("new AST for {k}");
+            ast::Int::fresh_const("pkg")
         });
     }
     for specs in smap.values() {
         for spec in specs {
             for r in (spec).reqs.keys() {
                 asts.entry(r.clone()).or_insert_with(|| {
-                    info!("new AST for {}", r);
-                    ast::Int::fresh_const(&ctx, "dep-pkg")
+                    info!("new AST for {r}");
+                    ast::Int::fresh_const("dep-pkg")
                 });
             }
         }
@@ -222,8 +219,8 @@ fn test_solve_simple_semver_example() {
                         );
                         opt.assert(
                             &k_ast
-                                .eq(&ast::Int::from_u64(&ctx, n as u64))
-                                .implies(&r_ast.ge(&ast::Int::from_u64(&ctx, low as u64))),
+                                ._eq(ast::Int::from_u64(n as u64))
+                                .implies(r_ast.ge(ast::Int::from_u64(low as u64))),
                         );
                     }
                 }
@@ -241,8 +238,8 @@ fn test_solve_simple_semver_example() {
                         );
                         opt.assert(
                             &k_ast
-                                .eq(&ast::Int::from_u64(&ctx, n as u64))
-                                .implies(&r_ast.le(&ast::Int::from_u64(&ctx, high as u64))),
+                                ._eq(ast::Int::from_u64(n as u64))
+                                .implies(r_ast.le(ast::Int::from_u64(high as u64))),
                         );
                     }
                 }
