@@ -1,32 +1,54 @@
+use std::path::Path;
 use std::{env, path::PathBuf};
 
+macro_rules! assert_one_of_features {
+    ($($feature:literal),*) => {{
+        let mut active_count = 0;
+        let mut active_feature = None;
+        $(
+            if cfg!(feature = $feature) {
+                active_count += 1;
+                active_feature = Some($feature);
+            }
+        )*
+        if active_count > 1 {
+            panic!("Only one of the features [{}] can be active at a time", stringify!($($feature),*));
+        }
+        active_feature
+    }};
+}
+
 fn main() {
-    // Don't need to specify this for these build configurations.
-    #[cfg(any(feature = "bundled", feature = "vcpkg"))]
-    let search_paths = vec![];
+    // Check that only one of the mutually exclusive features is active
+    let active_feature = assert_one_of_features!("bundled", "vcpkg", "gh-release");
 
-    #[cfg(feature = "bundled")]
-    build_bundled_z3();
+    println!("cargo:rerun-if-changed=build.rs");
 
-    #[cfg(not(any(feature = "bundled", feature = "vcpkg")))]
-    let search_paths = {
-        if let Ok(lib) = pkg_config::Config::new().probe("z3") {
-            lib.include_paths
-        } else {
-            vec![]
+    let (header, search_paths) = match active_feature {
+        Some("bundled") => {
+            build_bundled_z3();
+            (find_header_by_env(), vec![])
+        }
+        Some("gh-release") => (install_from_gh_release(), vec![]),
+        Some("vcpkg") => (find_library_header_by_vcpkg(), vec![]),
+        _ => {
+            let search_paths = if let Ok(lib) = pkg_config::Config::new().probe("z3") {
+                lib.include_paths
+            } else {
+                vec![]
+            };
+            println!("cargo:rerun-if-env-changed=Z3_LIBRARY_PATH_OVERRIDE");
+            if let Ok(lib_path) = env::var("Z3_LIBRARY_PATH_OVERRIDE") {
+                println!("cargo:rustc-link-search=native={lib_path}")
+            }
+            (find_header_by_env(), search_paths)
         }
     };
 
     #[cfg(feature = "deprecated-static-link-z3")]
-    println!("cargo:warning=The 'static-link-z3' feature is deprecated. Please use the 'bundled' feature.");
-
-    println!("cargo:rerun-if-changed=build.rs");
-
-    #[cfg(not(feature = "vcpkg"))]
-    let header = find_header_by_env();
-
-    #[cfg(feature = "vcpkg")]
-    let header = find_library_header_by_vcpkg();
+    println!(
+        "cargo:warning=The 'static-link-z3' feature is deprecated. Please use the 'bundled' feature."
+    );
 
     link_against_cxx_stdlib();
 
@@ -62,6 +84,159 @@ fn link_against_cxx_stdlib() {
     }
 }
 
+#[cfg(feature = "gh-release")]
+mod gh_release {
+    use std::path::Path;
+
+    use super::*;
+    use reqwest::blocking::{Client, ClientBuilder};
+    use reqwest::header::{AUTHORIZATION, HeaderMap};
+    use zip::ZipArchive;
+    use zip::read::root_dir_common_filter;
+
+    pub(super) fn install_from_gh_release() -> String {
+        let target_os = env::var("CARGO_CFG_TARGET_OS").unwrap();
+        let target_arch = env::var("CARGO_CFG_TARGET_ARCH").unwrap();
+        let (header, lib) = retrieve_gh_release_z3(&target_os, &target_arch);
+        println!(
+            "cargo:rustc-link-search=native={}",
+            lib.parent().unwrap().display()
+        );
+        if cfg!(target_os = "windows") {
+            println!("cargo:rustc-link-lib=static=libz3");
+        } else {
+            println!("cargo:rustc-link-lib=static=z3");
+        }
+        header.to_string_lossy().to_string()
+    }
+
+    fn retrieve_gh_release_z3(target_os: &str, target_arch: &str) -> (PathBuf, PathBuf) {
+        let arch = match target_arch {
+            "aarch64" => "arm64",
+            "x86_64" => "x64",
+            arch => {
+                panic!("Unsupported architecture: {}", arch);
+            }
+        };
+        let os = match target_os {
+            "windows" => "win",
+            "linux" => "glibc",
+            "macos" => "osx",
+            os => {
+                panic!("Unsupported OS: {}", os);
+            }
+        };
+        println!("cargo:rerun-if-env-changed=Z3_SYS_Z3_VERSION");
+        let z3_version = env::var("Z3_SYS_Z3_VERSION").unwrap_or("4.15.2".to_string());
+        let z3_dir = PathBuf::from(env::var("OUT_DIR").unwrap()).join(format!("z3-{z3_version}"));
+
+        if !z3_dir.exists() {
+            let client = get_github_client();
+
+            let url = get_release_asset_url(&client, &z3_version, os, arch);
+            if let Err(err) = download_unzip(&client, url, &z3_dir) {
+                println!("error: {err}");
+                panic!(
+                    "Could not get release asset for z3-{} with os={} and arch={}",
+                    z3_version, os, arch
+                );
+            };
+        } else {
+            println!("Found cached z3 at {}", z3_dir.display());
+        }
+
+        let header = z3_dir.join("include/z3.h");
+        let lib = if cfg!(target_os = "windows") {
+            z3_dir.join("bin/libz3.lib")
+        } else {
+            z3_dir.join("bin/libz3.a")
+        };
+
+        assert!(
+            header.exists(),
+            "could not find z3.h in downloaded archive at {}",
+            z3_dir.display()
+        );
+        assert!(
+            lib.exists(),
+            "could not find static libz3 in downloaded archive at {}",
+            z3_dir.display()
+        );
+
+        (header, lib)
+    }
+
+    fn download_unzip(client: &Client, url: String, dir: &Path) -> reqwest::Result<()> {
+        let response = client.get(url).send()?;
+        assert_eq!(response.status(), 200);
+        let ziplib = response.bytes()?;
+
+        println!("Downloaded {:0.2}MB", ziplib.len() as f64 / 1024.0 / 1024.0);
+
+        ZipArchive::new(std::io::Cursor::new(ziplib))
+            .unwrap()
+            .extract_unwrapped_root_dir(dir, root_dir_common_filter)
+            .expect("Failed to extract z3 release asset");
+        Ok(())
+    }
+
+    fn get_release_asset_url(
+        client: &Client,
+        z3_version: &str,
+        target_os: &str,
+        target_arch: &str,
+    ) -> String {
+        let release_url =
+            format!("https://api.github.com/repos/Z3Prover/z3/releases/tags/z3-{z3_version}");
+        let Ok(response) = client.get(release_url).send() else {
+            panic!("Could not find release for z3-{}", z3_version);
+        };
+
+        assert_eq!(response.status(), 200);
+
+        let release_json: serde_json::Value =
+            serde_json::from_str(&response.text().unwrap()).unwrap();
+
+        let assets = release_json.get("assets").unwrap().as_array().unwrap();
+
+        let Some(asset) = assets.iter().find(|a| {
+            let name = a.get("name").unwrap().as_str().unwrap();
+            name.contains(target_os)
+                && name.contains(target_arch)
+                && name.ends_with(".zip")
+                && name.starts_with("z3-")
+        }) else {
+            panic!(
+                "Could not find asset for z3-{} with os={} and arch={}",
+                z3_version, target_os, target_arch
+            );
+        };
+
+        asset
+            .get("browser_download_url")
+            .unwrap()
+            .as_str()
+            .unwrap()
+            .to_owned()
+    }
+
+    fn get_github_client() -> Client {
+        let client = ClientBuilder::new().user_agent("z3-sys");
+        let mut headers = HeaderMap::new();
+        if let Ok(val) = env::var("READ_ONLY_GITHUB_TOKEN") {
+            headers.insert(AUTHORIZATION, format!("Bearer {val}").parse().unwrap());
+        }
+        client.default_headers(headers).build().unwrap()
+    }
+}
+#[cfg(feature = "gh-release")]
+use gh_release::install_from_gh_release;
+
+#[cfg(not(feature = "gh-release"))]
+fn install_from_gh_release() -> String {
+    unreachable!()
+}
+
 #[cfg(feature = "vcpkg")]
 fn find_library_header_by_vcpkg() -> String {
     let lib = vcpkg::Config::new()
@@ -81,10 +256,24 @@ fn find_library_header_by_vcpkg() -> String {
 }
 
 #[cfg(not(feature = "vcpkg"))]
+fn find_library_header_by_vcpkg() -> String {
+    unreachable!()
+}
+
 fn find_header_by_env() -> String {
     const Z3_HEADER_VAR: &str = "Z3_SYS_Z3_HEADER";
     let header = if cfg!(feature = "bundled") {
-        "z3/src/api/z3.h".to_string()
+        env::var("Z3_SYS_BUNDLED_DIR_OVERRIDE")
+            .map(|a| {
+                Path::new(&a)
+                    .join("src")
+                    .join("api")
+                    .join("z3.h")
+                    .to_str()
+                    .unwrap()
+                    .to_string()
+            })
+            .unwrap_or("z3/src/api/z3.h".to_string())
     } else if let Ok(header_path) = env::var(Z3_HEADER_VAR) {
         header_path
     } else {
@@ -109,13 +298,17 @@ fn generate_binding(header: &str, search_paths: &[PathBuf]) {
         "sort_kind",
         "symbol_kind",
     ] {
+        #[allow(unused_mut)]
         let mut enum_bindings = bindgen::Builder::default()
             .header(header)
-            .parse_callbacks(Box::new(bindgen::CargoCallbacks::new()))
             .generate_comments(false)
             .rustified_enum(format!("Z3_{x}"))
             .allowlist_type(format!("Z3_{x}"))
             .clang_args(search_paths.iter().map(|p| format!("-I{}", p.display())));
+        // Deactivate bindgen cargo-rerun-if generation for gh-release (unnecessary)
+        #[cfg(not(feature = "gh-release"))]
+        let mut enum_bindings =
+            enum_bindings.parse_callbacks(Box::new(bindgen::CargoCallbacks::new()));
         if env::var("TARGET").unwrap() == "wasm32-unknown-emscripten" {
             enum_bindings = enum_bindings.clang_arg(format!(
                 "--sysroot={}/upstream/emscripten/cache/sysroot",
@@ -133,10 +326,10 @@ fn generate_binding(header: &str, search_paths: &[PathBuf]) {
 /// Build z3 with bundled source codes.
 #[cfg(feature = "bundled")]
 fn build_bundled_z3() {
-    let mut cfg = cmake::Config::new("z3");
-    cfg
-        // Don't build `libz3.so`, build `libz3.a` instead.
-        .define("Z3_BUILD_LIBZ3_SHARED", "false")
+    let bundled_path = env::var("Z3_SYS_BUNDLED_DIR_OVERRIDE").unwrap_or("z3".to_string());
+    let mut cfg = cmake::Config::new(bundled_path);
+    // Don't build `libz3.so`, build `libz3.a` instead.
+    cfg.define("Z3_BUILD_LIBZ3_SHARED", "false")
         // Don't build the Z3 repl.
         .define("Z3_BUILD_EXECUTABLE", "false")
         // Don't build the tests.
@@ -151,6 +344,10 @@ fn build_bundled_z3() {
         cfg.cxxflag("-DWIN32");
         cfg.cxxflag("-D_WINDOWS");
         cfg.define("CMAKE_MSVC_RUNTIME_LIBRARY", "MultiThreadedDLL");
+    } else if env::var("TARGET").unwrap().starts_with("wasm") {
+        // for wasm targets, ensure we allow exceptions
+        // because z3 has some exceptions
+        cfg.no_default_flags(true).cxxflag("-fexceptions");
     }
 
     let dst = cfg.build();
@@ -182,4 +379,8 @@ fn build_bundled_z3() {
     } else {
         println!("cargo:rustc-link-lib=static=z3");
     }
+}
+#[cfg(not(feature = "bundled"))]
+fn build_bundled_z3() {
+    unreachable!()
 }
