@@ -1,34 +1,35 @@
 use std::convert::TryInto;
 use std::ffi::CStr;
 use std::fmt;
-
 use z3_sys::*;
 
-use crate::{ast, ast::Ast, Context, FuncDecl, Sort, Symbol};
-
-impl<'ctx> FuncDecl<'ctx> {
-    pub(crate) unsafe fn wrap(ctx: &'ctx Context, z3_func_decl: Z3_func_decl) -> Self {
-        Z3_inc_ref(ctx.z3_ctx, Z3_func_decl_to_ast(ctx.z3_ctx, z3_func_decl));
-        Self { ctx, z3_func_decl }
+use crate::{Context, FuncDecl, Sort, Symbol, ast, ast::Ast};
+impl FuncDecl {
+    pub(crate) unsafe fn wrap(ctx: &Context, z3_func_decl: Z3_func_decl) -> Self {
+        unsafe {
+            Z3_inc_ref(
+                ctx.z3_ctx.0,
+                Z3_func_decl_to_ast(ctx.z3_ctx.0, z3_func_decl),
+            );
+        }
+        Self {
+            ctx: ctx.clone(),
+            z3_func_decl,
+        }
     }
 
-    pub fn new<S: Into<Symbol>>(
-        ctx: &'ctx Context,
-        name: S,
-        domain: &[&Sort<'ctx>],
-        range: &Sort<'ctx>,
-    ) -> Self {
+    pub fn new<S: Into<Symbol>>(name: S, domain: &[&Sort], range: &Sort) -> Self {
+        let ctx = &Context::thread_local();
         assert!(domain.iter().all(|s| s.ctx.z3_ctx == ctx.z3_ctx));
         assert_eq!(ctx.z3_ctx, range.ctx.z3_ctx);
 
         let domain: Vec<_> = domain.iter().map(|s| s.z3_sort).collect();
-
         unsafe {
             Self::wrap(
                 ctx,
                 Z3_mk_func_decl(
-                    ctx.z3_ctx,
-                    name.into().as_z3_symbol(ctx),
+                    ctx.z3_ctx.0,
+                    name.into().as_z3_symbol(),
                     domain.len().try_into().unwrap(),
                     domain.as_ptr(),
                     range.z3_sort,
@@ -43,31 +44,28 @@ impl<'ctx> FuncDecl<'ctx> {
     ///
     /// ```
     /// # use z3::{Config, Context, FuncDecl, Solver, Sort, Symbol};
-    /// # let cfg = Config::new();
-    /// # let ctx = Context::new(&cfg);
     /// let f = FuncDecl::new(
-    ///     &ctx,
     ///     "f",
-    ///     &[&Sort::int(&ctx), &Sort::real(&ctx)],
-    ///     &Sort::int(&ctx));
+    ///     &[&Sort::int(), &Sort::real()],
+    ///     &Sort::int());
     /// assert_eq!(f.arity(), 2);
     /// ```
     pub fn arity(&self) -> usize {
-        unsafe { Z3_get_arity(self.ctx.z3_ctx, self.z3_func_decl) as usize }
+        unsafe { Z3_get_arity(self.ctx.z3_ctx.0, self.z3_func_decl) as usize }
     }
 
     /// Create a constant (if `args` has length 0) or function application (otherwise).
     ///
     /// Note that `args` should have the types corresponding to the `domain` of the `FuncDecl`.
-    pub fn apply(&self, args: &[&dyn ast::Ast<'ctx>]) -> ast::Dynamic<'ctx> {
+    pub fn apply(&self, args: &[&dyn ast::Ast]) -> ast::Dynamic {
         assert!(args.iter().all(|s| s.get_ctx().z3_ctx == self.ctx.z3_ctx));
 
         let args: Vec<_> = args.iter().map(|a| a.get_z3_ast()).collect();
 
         unsafe {
-            ast::Dynamic::wrap(self.ctx, {
+            ast::Dynamic::wrap(&self.ctx, {
                 Z3_mk_app(
-                    self.ctx.z3_ctx,
+                    self.ctx.z3_ctx.0,
                     self.z3_func_decl,
                     args.len().try_into().unwrap(),
                     args.as_ptr(),
@@ -78,7 +76,7 @@ impl<'ctx> FuncDecl<'ctx> {
 
     /// Return the `DeclKind` of this `FuncDecl`.
     pub fn kind(&self) -> DeclKind {
-        unsafe { Z3_get_decl_kind(self.ctx.z3_ctx, self.z3_func_decl) }
+        unsafe { Z3_get_decl_kind(self.ctx.z3_ctx.0, self.z3_func_decl) }
     }
 
     /// Return the name of this `FuncDecl`.
@@ -87,7 +85,7 @@ impl<'ctx> FuncDecl<'ctx> {
     /// the `Symbol`.
     pub fn name(&self) -> String {
         unsafe {
-            let z3_ctx = self.ctx.z3_ctx;
+            let z3_ctx = self.ctx.z3_ctx.0;
             let symbol = Z3_get_decl_name(z3_ctx, self.z3_func_decl);
             match Z3_get_symbol_kind(z3_ctx, symbol) {
                 SymbolKind::String => CStr::from_ptr(Z3_get_symbol_string(z3_ctx, symbol))
@@ -99,9 +97,9 @@ impl<'ctx> FuncDecl<'ctx> {
     }
 }
 
-impl<'ctx> fmt::Display for FuncDecl<'ctx> {
+impl fmt::Display for FuncDecl {
     fn fmt(&self, f: &mut fmt::Formatter) -> Result<(), fmt::Error> {
-        let p = unsafe { Z3_func_decl_to_string(self.ctx.z3_ctx, self.z3_func_decl) };
+        let p = unsafe { Z3_func_decl_to_string(self.ctx.z3_ctx.0, self.z3_func_decl) };
         if p.is_null() {
             return Result::Err(fmt::Error);
         }
@@ -112,18 +110,18 @@ impl<'ctx> fmt::Display for FuncDecl<'ctx> {
     }
 }
 
-impl<'ctx> fmt::Debug for FuncDecl<'ctx> {
+impl fmt::Debug for FuncDecl {
     fn fmt(&self, f: &mut fmt::Formatter) -> Result<(), fmt::Error> {
         <Self as fmt::Display>::fmt(self, f)
     }
 }
 
-impl<'ctx> Drop for FuncDecl<'ctx> {
+impl Drop for FuncDecl {
     fn drop(&mut self) {
         unsafe {
             Z3_dec_ref(
-                self.ctx.z3_ctx,
-                Z3_func_decl_to_ast(self.ctx.z3_ctx, self.z3_func_decl),
+                self.ctx.z3_ctx.0,
+                Z3_func_decl_to_ast(self.ctx.z3_ctx.0, self.z3_func_decl),
             );
         }
     }
