@@ -1,31 +1,35 @@
 use std::fmt;
-
 use z3_sys::*;
 
 use crate::{
-    ast::{Ast, Dynamic},
     Context, FuncEntry,
+    ast::{Ast, Dynamic},
 };
 
-impl<'ctx> FuncEntry<'ctx> {
-    pub(crate) unsafe fn wrap(ctx: &'ctx Context, z3_func_entry: Z3_func_entry) -> Self {
-        Z3_func_entry_inc_ref(ctx.z3_ctx, z3_func_entry);
-        Self { ctx, z3_func_entry }
+impl FuncEntry {
+    pub(crate) unsafe fn wrap(ctx: &Context, z3_func_entry: Z3_func_entry) -> Self {
+        unsafe {
+            Z3_func_entry_inc_ref(ctx.z3_ctx.0, z3_func_entry);
+        }
+        Self {
+            ctx: ctx.clone(),
+            z3_func_entry,
+        }
     }
 
     /// Returns the value of the function.
     pub fn get_value(&self) -> Dynamic {
         unsafe {
             Dynamic::wrap(
-                self.ctx,
-                Z3_func_entry_get_value(self.ctx.z3_ctx, self.z3_func_entry),
+                &self.ctx,
+                Z3_func_entry_get_value(self.ctx.z3_ctx.0, self.z3_func_entry),
             )
         }
     }
 
     /// Returns the number of arguments in the function entry.
     pub fn get_num_args(&self) -> u32 {
-        unsafe { Z3_func_entry_get_num_args(self.ctx.z3_ctx, self.z3_func_entry) }
+        unsafe { Z3_func_entry_get_num_args(self.ctx.z3_ctx.0, self.z3_func_entry) }
     }
 
     /// Returns the arguments of the function entry.
@@ -33,15 +37,15 @@ impl<'ctx> FuncEntry<'ctx> {
         (0..self.get_num_args())
             .map(|i| unsafe {
                 Dynamic::wrap(
-                    self.ctx,
-                    Z3_func_entry_get_arg(self.ctx.z3_ctx, self.z3_func_entry, i),
+                    &self.ctx,
+                    Z3_func_entry_get_arg(self.ctx.z3_ctx.0, self.z3_func_entry, i),
                 )
             })
             .collect()
     }
 }
 
-impl<'ctx> fmt::Display for FuncEntry<'ctx> {
+impl fmt::Display for FuncEntry {
     fn fmt(&self, f: &mut fmt::Formatter) -> Result<(), fmt::Error> {
         write!(f, "[")?;
         self.get_args()
@@ -52,16 +56,16 @@ impl<'ctx> fmt::Display for FuncEntry<'ctx> {
     }
 }
 
-impl<'ctx> fmt::Debug for FuncEntry<'ctx> {
+impl fmt::Debug for FuncEntry {
     fn fmt(&self, f: &mut fmt::Formatter) -> Result<(), fmt::Error> {
         <Self as fmt::Display>::fmt(self, f)
     }
 }
 
-impl<'ctx> Drop for FuncEntry<'ctx> {
+impl Drop for FuncEntry {
     fn drop(&mut self) {
         unsafe {
-            Z3_func_entry_dec_ref(self.ctx.z3_ctx, self.z3_func_entry);
+            Z3_func_entry_dec_ref(self.ctx.z3_ctx.0, self.z3_func_entry);
         }
     }
 }
