@@ -30,12 +30,16 @@ mod sort;
 mod statistics;
 mod symbol;
 mod tactic;
+mod translate;
 mod version;
 
 pub use crate::params::{get_global_param, reset_all_global_params, set_global_param};
 pub use crate::statistics::{StatisticsEntry, StatisticsValue};
-pub use crate::version::{full_version, version, Version};
-
+pub use crate::translate::Translate;
+pub use crate::translate::synchronization::*;
+pub use crate::version::{Version, full_version, version};
+pub use context::Context;
+pub use datatype_builder::DatatypeAccessor;
 /// Configuration used to initialize [logical contexts](Context).
 ///
 /// # See also:
@@ -45,31 +49,6 @@ pub use crate::version::{full_version, version, Version};
 pub struct Config {
     kvs: Vec<(CString, CString)>,
     z3_cfg: Z3_config,
-}
-
-/// Manager of all other Z3 objects, global configuration options, etc.
-///
-/// An application may use multiple Z3 contexts. Objects created in one context
-/// cannot be used in another one. However, several objects may be "translated" from
-/// one context to another. It is not safe to access Z3 objects from multiple threads.
-///
-/// # Examples:
-///
-/// Creating a context with the default configuration:
-///
-/// ```
-/// use z3::{Config, Context};
-/// let cfg = Config::new();
-/// let ctx = Context::new(&cfg);
-/// ```
-///
-/// # See also:
-///
-/// - [`Config`]
-/// - [`Context::new()`]
-#[derive(PartialEq, Eq, Debug)]
-pub struct Context {
-    z3_ctx: Z3_context,
 }
 
 /// Handle that can be used to interrupt a computation from another thread.
@@ -95,16 +74,16 @@ pub enum Symbol {
 //
 // Note for in-crate users: Never construct a `Sort` directly; only use
 // `Sort::new()` which handles Z3 refcounting properly.
-pub struct Sort<'ctx> {
-    ctx: &'ctx Context,
+pub struct Sort {
+    ctx: Context,
     z3_sort: Z3_sort,
 }
 
 /// A struct to represent when two sorts are of different types.
 #[derive(Debug)]
-pub struct SortDiffers<'ctx> {
-    left: Sort<'ctx>,
-    right: Sort<'ctx>,
+pub struct SortDiffers {
+    left: Sort,
+    right: Sort,
 }
 
 /// A struct to represent when an ast is not a function application.
@@ -117,8 +96,8 @@ pub struct IsNotApp {
 //
 // Note for in-crate users: Never construct a `Solver` directly; only use
 // `Solver::new()` which handles Z3 refcounting properly.
-pub struct Solver<'ctx> {
-    ctx: &'ctx Context,
+pub struct Solver {
+    ctx: Context,
     z3_slv: Z3_solver,
 }
 
@@ -126,8 +105,8 @@ pub struct Solver<'ctx> {
 //
 // Note for in-crate users: Never construct a `Model` directly; only use
 // `Model::new()` which handles Z3 refcounting properly.
-pub struct Model<'ctx> {
-    ctx: &'ctx Context,
+pub struct Model {
+    ctx: Context,
     z3_mdl: Z3_model,
 }
 
@@ -135,8 +114,8 @@ pub struct Model<'ctx> {
 //
 // Note for in-crate users: Never construct an `Optimize` directly; only use
 // `Optimize::new()` which handles Z3 refcounting properly.
-pub struct Optimize<'ctx> {
-    ctx: &'ctx Context,
+pub struct Optimize {
+    ctx: Context,
     z3_opt: Z3_optimize,
 }
 
@@ -152,22 +131,22 @@ pub struct Optimize<'ctx> {
 //
 // Note for in-crate users: Never construct a `FuncDecl` directly; only use
 // `FuncDecl::new()` which handles Z3 refcounting properly.
-pub struct FuncDecl<'ctx> {
-    ctx: &'ctx Context,
+pub struct FuncDecl {
+    ctx: Context,
     z3_func_decl: Z3_func_decl,
 }
 
 /// Stores the interpretation of a function in a Z3 model.
 /// <https://z3prover.github.io/api/html/classz3py_1_1_func_interp.html>
-pub struct FuncInterp<'ctx> {
-    ctx: &'ctx Context,
+pub struct FuncInterp {
+    ctx: Context,
     z3_func_interp: Z3_func_interp,
 }
 
 /// Store the value of the interpretation of a function in a particular point.
 /// <https://z3prover.github.io/api/html/classz3py_1_1_func_entry.html>
-pub struct FuncEntry<'ctx> {
-    ctx: &'ctx Context,
+pub struct FuncEntry {
+    ctx: Context,
     z3_func_entry: Z3_func_entry,
 }
 
@@ -184,8 +163,8 @@ pub struct FuncEntry<'ctx> {
 /// - [`RecFuncDecl::add_def`]
 // Note for in-crate users: Never construct a `RecFuncDecl` directly; only use
 // `RecFuncDecl::new()` which handles Z3 refcounting properly.
-pub struct RecFuncDecl<'ctx> {
-    ctx: &'ctx Context,
+pub struct RecFuncDecl {
+    ctx: Context,
     z3_func_decl: Z3_func_decl,
 }
 
@@ -195,67 +174,60 @@ pub use z3_sys::DeclKind;
 ///
 /// Example:
 /// ```
-/// # use z3::{ast::Int, Config, Context, DatatypeAccessor, DatatypeBuilder, SatResult, Solver, Sort, ast::{Ast, Datatype}};
+/// # use z3::{ Config, Context, DatatypeAccessor, DatatypeBuilder, SatResult, Solver, Sort, ast::{Ast, Datatype, Int}};
 /// # let cfg = Config::new();
 /// # let ctx = Context::new(&cfg);
-/// # let solver = Solver::new(&ctx);
+/// # let solver = Solver::new();
 /// // Like Rust's Option<int> type
-/// let option_int = DatatypeBuilder::new(&ctx, "OptionInt")
+/// let option_int = DatatypeBuilder::new("OptionInt")
 /// .variant("None", vec![])
 /// .variant(
 ///     "Some",
-///     vec![("value", DatatypeAccessor::Sort(Sort::int(&ctx)))],
+///     vec![("value", DatatypeAccessor::Sort(Sort::int()))],
 /// )
 /// .finish();
 ///
 /// // Assert x.is_none()
-/// let x = Datatype::new_const(&ctx, "x", &option_int.sort);
+/// let x = Datatype::new_const("x", &option_int.sort);
 /// solver.assert(&option_int.variants[0].tester.apply(&[&x]).as_bool().unwrap());
 ///
 /// // Assert y == Some(3)
-/// let y = Datatype::new_const(&ctx, "y", &option_int.sort);
-/// let value = option_int.variants[1].constructor.apply(&[&Int::from_i64(&ctx, 3)]);
+/// let y = Datatype::new_const("y", &option_int.sort);
+/// let value = option_int.variants[1].constructor.apply(&[&Int::from_i64(3)]);
 /// solver.assert(&y._eq(&value.as_datatype().unwrap()));
 ///
 /// assert_eq!(solver.check(), SatResult::Sat);
-/// let model = solver.get_model().unwrap();;
+/// let model = solver.get_model().unwrap();
 ///
 /// // Get the value out of Some(3)
 /// let ast = option_int.variants[1].accessors[0].apply(&[&y]);
 /// assert_eq!(3, model.eval(&ast.as_int().unwrap(), true).unwrap().as_i64().unwrap());
 /// ```
 #[derive(Debug)]
-pub struct DatatypeBuilder<'ctx> {
-    ctx: &'ctx Context,
+pub struct DatatypeBuilder {
+    ctx: Context,
     name: Symbol,
-    constructors: Vec<(String, Vec<(String, DatatypeAccessor<'ctx>)>)>,
-}
-
-/// Wrapper which can point to a sort (by value) or to a custom datatype (by name).
-#[derive(Debug)]
-pub enum DatatypeAccessor<'ctx> {
-    Sort(Sort<'ctx>),
-    Datatype(Symbol),
+    constructors: Vec<(String, Vec<(String, DatatypeAccessor)>)>,
 }
 
 /// Inner variant for a custom [datatype sort](DatatypeSort).
 #[derive(Debug)]
-pub struct DatatypeVariant<'ctx> {
-    pub constructor: FuncDecl<'ctx>,
-    pub tester: FuncDecl<'ctx>,
-    pub accessors: Vec<FuncDecl<'ctx>>,
+pub struct DatatypeVariant {
+    pub constructor: FuncDecl,
+    pub tester: FuncDecl,
+    pub accessors: Vec<FuncDecl>,
 }
 
 /// A custom datatype sort.
 #[derive(Debug)]
-pub struct DatatypeSort<'ctx> {
-    pub sort: Sort<'ctx>,
-    pub variants: Vec<DatatypeVariant<'ctx>>,
+pub struct DatatypeSort {
+    pub sort: Sort,
+    pub variants: Vec<DatatypeVariant>,
 }
 
 /// Parameter set used to configure many components (simplifiers, tactics, solvers, etc).
-pub struct Params<'ctx> {
-    ctx: &'ctx Context,
+pub struct Params {
+    ctx: Context,
     z3_params: Z3_params,
 }
 
@@ -271,15 +243,15 @@ pub enum SatResult {
 }
 
 /// A pattern for quantifier instantiation, used to guide quantifier instantiation.
-pub struct Pattern<'ctx> {
-    ctx: &'ctx Context,
+pub struct Pattern {
+    ctx: Context,
     z3_pattern: Z3_pattern,
 }
 
 /// Collection of subgoals resulting from applying of a tactic to a goal.
 #[derive(Clone, Debug)]
-pub struct ApplyResult<'ctx> {
-    ctx: &'ctx Context,
+pub struct ApplyResult {
+    ctx: Context,
     z3_apply_result: Z3_apply_result,
 }
 
@@ -301,14 +273,14 @@ pub struct ApplyResult<'ctx> {
 ///
 /// Finally, a solver utilizing a tactic can be created via
 /// [`Tactic::solver()`].
-pub struct Tactic<'ctx> {
-    ctx: &'ctx Context,
+pub struct Tactic {
+    ctx: Context,
     z3_tactic: Z3_tactic,
 }
 
 /// Set of formulas that can be solved and/or transformed using tactics and solvers.
-pub struct Goal<'ctx> {
-    ctx: &'ctx Context,
+pub struct Goal {
+    ctx: Context,
     z3_goal: Z3_goal,
 }
 
@@ -318,8 +290,8 @@ pub struct Goal<'ctx> {
 ///
 /// Z3 provides a variety of probes, which can be queried via
 /// [`Probe::list_all()`].
-pub struct Probe<'ctx> {
-    ctx: &'ctx Context,
+pub struct Probe {
+    ctx: Context,
     z3_probe: Z3_probe,
 }
 
@@ -329,7 +301,38 @@ pub struct Probe<'ctx> {
 ///
 /// - [`Optimize::get_statistics()`]
 /// - [`Solver::get_statistics()`]
-pub struct Statistics<'ctx> {
-    ctx: &'ctx Context,
+pub struct Statistics {
+    ctx: Context,
     z3_stats: Z3_stats,
+}
+
+/// Runs the provided callback with all inner Z3 calls using the provided [`Context`].
+///
+/// Requires that the closure and return type be [`Send`] and [`Sync`] to prevent
+/// mixing Z3 objects belonging to multiple [`Context`]s. If you need to move Z3 data
+/// into or out of the closure, use [`PrepareSynchronized::synchronized()`].
+/// # See also
+///
+/// [`with_z3_config`]
+#[deprecated = "Use `with_z3_config` instead, which constructs a Context from a Config"]
+pub fn with_z3_context<T: Fn() -> R + Send + Sync, R: Send + Sync>(
+    ctx: &Context,
+    callback: T,
+) -> R {
+    let old = Context::thread_local();
+    Context::set_thread_local(ctx);
+    let res = callback();
+    Context::set_thread_local(&old);
+    res
+}
+
+/// Runs the provided callback with all inner Z3 calls using a [`Context`] derived
+/// from the provided [`Config`].
+///
+/// Requires that the closure and return type be [`Send`] and [`Sync`] to prevent
+/// mixing Z3 objects belonging to multiple [`Context`]s. If you need to move Z3 data
+/// into or out of the closure, use [`PrepareSynchronized::synchronized()`].
+#[allow(deprecated)]
+pub fn with_z3_config<T: Fn() -> R + Send + Sync, R: Send + Sync>(cfg: &Config, callback: T) -> R {
+    with_z3_context(&Context::new(cfg), callback)
 }
