@@ -1,84 +1,79 @@
 use std::ffi::CStr;
 use std::fmt;
-
 use z3_sys::*;
 
-use crate::{ast::Ast, Context, FuncDecl, FuncInterp, Model, Optimize, Solver};
+use crate::{Context, FuncDecl, FuncInterp, Model, Optimize, Solver, Translate, ast::Ast};
 
-impl<'ctx> Model<'ctx> {
-    unsafe fn wrap(ctx: &'ctx Context, z3_mdl: Z3_model) -> Model<'ctx> {
-        Z3_model_inc_ref(ctx.z3_ctx, z3_mdl);
-        Model { ctx, z3_mdl }
+impl Model {
+    unsafe fn wrap(ctx: &Context, z3_mdl: Z3_model) -> Model {
+        unsafe {
+            Z3_model_inc_ref(ctx.z3_ctx.0, z3_mdl);
+        }
+        Model {
+            ctx: ctx.clone(),
+            z3_mdl,
+        }
     }
 
-    pub fn of_solver(slv: &Solver<'ctx>) -> Option<Model<'ctx>> {
+    pub fn of_solver(slv: &Solver) -> Option<Model> {
         unsafe {
-            let m = Z3_solver_get_model(slv.ctx.z3_ctx, slv.z3_slv);
+            let m = Z3_solver_get_model(slv.ctx.z3_ctx.0, slv.z3_slv);
             if m.is_null() {
                 None
             } else {
-                Some(Self::wrap(slv.ctx, m))
+                Some(Self::wrap(&slv.ctx, m))
             }
         }
     }
 
-    pub fn of_optimize(opt: &Optimize<'ctx>) -> Option<Model<'ctx>> {
+    pub fn of_optimize(opt: &Optimize) -> Option<Model> {
         unsafe {
-            let m = Z3_optimize_get_model(opt.ctx.z3_ctx, opt.z3_opt);
+            let m = Z3_optimize_get_model(opt.ctx.z3_ctx.0, opt.z3_opt);
             if m.is_null() {
                 None
             } else {
-                Some(Self::wrap(opt.ctx, m))
+                Some(Self::wrap(&opt.ctx, m))
             }
-        }
-    }
-
-    /// Translate model to context `dest`
-    pub fn translate<'dest_ctx>(&self, dest: &'dest_ctx Context) -> Model<'dest_ctx> {
-        unsafe {
-            Model::wrap(
-                dest,
-                Z3_model_translate(self.ctx.z3_ctx, self.z3_mdl, dest.z3_ctx),
-            )
         }
     }
 
     /// Returns the interpretation of the given `ast` in the `Model`
     /// Returns `None` if there is no interpretation in the `Model`
-    pub fn get_const_interp<T: Ast<'ctx>>(&self, ast: &T) -> Option<T> {
+    pub fn get_const_interp<T: Ast>(&self, ast: &T) -> Option<T> {
         let func = ast.safe_decl().ok()?;
 
         let ret =
-            unsafe { Z3_model_get_const_interp(self.ctx.z3_ctx, self.z3_mdl, func.z3_func_decl) };
+            unsafe { Z3_model_get_const_interp(self.ctx.z3_ctx.0, self.z3_mdl, func.z3_func_decl) };
         if ret.is_null() {
             None
         } else {
-            Some(unsafe { T::wrap(self.ctx, ret) })
+            Some(unsafe { T::wrap(&self.ctx, ret) })
         }
     }
 
     /// Returns the interpretation of the given `f` in the `Model`
     /// Returns `None` if there is no interpretation in the `Model`
-    pub fn get_func_interp(&self, f: &FuncDecl) -> Option<FuncInterp<'ctx>> {
+    pub fn get_func_interp(&self, f: &FuncDecl) -> Option<FuncInterp> {
         if f.arity() == 0 {
-            let ret =
-                unsafe { Z3_model_get_const_interp(self.ctx.z3_ctx, self.z3_mdl, f.z3_func_decl) };
+            let ret = unsafe {
+                Z3_model_get_const_interp(self.ctx.z3_ctx.0, self.z3_mdl, f.z3_func_decl)
+            };
             if ret.is_null() {
                 None
             } else {
                 let sort_kind = unsafe {
                     Z3_get_sort_kind(
-                        self.ctx.z3_ctx,
-                        Z3_get_range(self.ctx.z3_ctx, f.z3_func_decl),
+                        self.ctx.z3_ctx.0,
+                        Z3_get_range(self.ctx.z3_ctx.0, f.z3_func_decl),
                     )
                 };
                 match sort_kind {
                     SortKind::Array => {
-                        if unsafe { Z3_is_as_array(self.ctx.z3_ctx, ret) } {
+                        if unsafe { Z3_is_as_array(self.ctx.z3_ctx.0, ret) } {
                             let fd = unsafe {
                                 FuncDecl::wrap(
-                                    self.ctx,
-                                    Z3_get_as_array_func_decl(self.ctx.z3_ctx, ret),
+                                    &self.ctx,
+                                    Z3_get_as_array_func_decl(self.ctx.z3_ctx.0, ret),
                                 )
                             };
                             self.get_func_interp(&fd)
@@ -91,24 +86,24 @@ impl<'ctx> Model<'ctx> {
             }
         } else {
             let ret =
-                unsafe { Z3_model_get_func_interp(self.ctx.z3_ctx, self.z3_mdl, f.z3_func_decl) };
+                unsafe { Z3_model_get_func_interp(self.ctx.z3_ctx.0, self.z3_mdl, f.z3_func_decl) };
             if ret.is_null() {
                 None
             } else {
-                Some(unsafe { FuncInterp::wrap(self.ctx, ret) })
+                Some(unsafe { FuncInterp::wrap(&self.ctx, ret) })
             }
         }
     }
 
     pub fn eval<T>(&self, ast: &T, model_completion: bool) -> Option<T>
     where
-        T: Ast<'ctx>,
+        T: Ast,
     {
         let mut tmp: Z3_ast = ast.get_z3_ast();
         let res = {
             unsafe {
                 Z3_model_eval(
-                    self.ctx.z3_ctx,
+                    self.ctx.z3_ctx.0,
                     self.z3_mdl,
                     ast.get_z3_ast(),
                     model_completion,
@@ -117,7 +112,7 @@ impl<'ctx> Model<'ctx> {
             }
         };
         if res {
-            Some(unsafe { T::wrap(self.ctx, tmp) })
+            Some(unsafe { T::wrap(&self.ctx, tmp) })
         } else {
             None
         }
@@ -125,19 +120,19 @@ impl<'ctx> Model<'ctx> {
 
     fn len(&self) -> u32 {
         unsafe {
-            Z3_model_get_num_consts(self.ctx.z3_ctx, self.z3_mdl)
-                + Z3_model_get_num_funcs(self.ctx.z3_ctx, self.z3_mdl)
+            Z3_model_get_num_consts(self.ctx.z3_ctx.0, self.z3_mdl)
+                + Z3_model_get_num_funcs(self.ctx.z3_ctx.0, self.z3_mdl)
         }
     }
 
-    pub fn iter(&'ctx self) -> ModelIter<'ctx> {
+    pub fn iter<'a>(&'a self) -> ModelIter<'a> {
         self.into_iter()
     }
 }
 
-impl<'ctx> fmt::Display for Model<'ctx> {
+impl fmt::Display for Model {
     fn fmt(&self, f: &mut fmt::Formatter) -> Result<(), fmt::Error> {
-        let p = unsafe { Z3_model_to_string(self.ctx.z3_ctx, self.z3_mdl) };
+        let p = unsafe { Z3_model_to_string(self.ctx.z3_ctx.0, self.z3_mdl) };
         if p.is_null() {
             return Result::Err(fmt::Error);
         }
@@ -148,29 +143,29 @@ impl<'ctx> fmt::Display for Model<'ctx> {
     }
 }
 
-impl<'ctx> fmt::Debug for Model<'ctx> {
+impl fmt::Debug for Model {
     fn fmt(&self, f: &mut fmt::Formatter) -> Result<(), fmt::Error> {
         <Self as fmt::Display>::fmt(self, f)
     }
 }
 
-impl<'ctx> Drop for Model<'ctx> {
+impl Drop for Model {
     fn drop(&mut self) {
-        unsafe { Z3_model_dec_ref(self.ctx.z3_ctx, self.z3_mdl) };
+        unsafe { Z3_model_dec_ref(self.ctx.z3_ctx.0, self.z3_mdl) };
     }
 }
 
 #[derive(Debug)]
 /// <https://z3prover.github.io/api/html/classz3py_1_1_model_ref.html#a7890b7c9bc70cf2a26a343c22d2c8367>
-pub struct ModelIter<'ctx> {
-    model: &'ctx Model<'ctx>,
+pub struct ModelIter<'a> {
+    model: &'a Model,
     idx: u32,
     len: u32,
 }
 
-impl<'ctx> IntoIterator for &'ctx Model<'ctx> {
-    type Item = FuncDecl<'ctx>;
-    type IntoIter = ModelIter<'ctx>;
+impl<'a> IntoIterator for &'a Model {
+    type Item = FuncDecl;
+    type IntoIter = ModelIter<'a>;
 
     fn into_iter(self) -> Self::IntoIter {
         ModelIter {
@@ -181,31 +176,31 @@ impl<'ctx> IntoIterator for &'ctx Model<'ctx> {
     }
 }
 
-impl<'ctx> Iterator for ModelIter<'ctx> {
-    type Item = FuncDecl<'ctx>;
+impl Iterator for ModelIter<'_> {
+    type Item = FuncDecl;
 
     fn next(&mut self) -> Option<Self::Item> {
         if self.idx >= self.len {
             None
         } else {
             let num_consts =
-                unsafe { Z3_model_get_num_consts(self.model.ctx.z3_ctx, self.model.z3_mdl) };
+                unsafe { Z3_model_get_num_consts(self.model.ctx.z3_ctx.0, self.model.z3_mdl) };
             if self.idx < num_consts {
                 let const_decl = unsafe {
-                    Z3_model_get_const_decl(self.model.ctx.z3_ctx, self.model.z3_mdl, self.idx)
+                    Z3_model_get_const_decl(self.model.ctx.z3_ctx.0, self.model.z3_mdl, self.idx)
                 };
                 self.idx += 1;
-                Some(unsafe { FuncDecl::wrap(self.model.ctx, const_decl) })
+                Some(unsafe { FuncDecl::wrap(&self.model.ctx, const_decl) })
             } else {
                 let func_decl = unsafe {
                     Z3_model_get_func_decl(
-                        self.model.ctx.z3_ctx,
+                        self.model.ctx.z3_ctx.0,
                         self.model.z3_mdl,
                         self.idx - num_consts,
                     )
                 };
                 self.idx += 1;
-                Some(unsafe { FuncDecl::wrap(self.model.ctx, func_decl) })
+                Some(unsafe { FuncDecl::wrap(&self.model.ctx, func_decl) })
             }
         }
     }
@@ -216,13 +211,22 @@ impl<'ctx> Iterator for ModelIter<'ctx> {
     }
 }
 
+unsafe impl Translate for Model {
+    fn translate(&self, dest: &Context) -> Model {
+        unsafe {
+            Model::wrap(
+                dest,
+                Z3_model_translate(self.ctx.z3_ctx.0, self.z3_mdl, dest.z3_ctx.0),
+            )
+        }
+    }
+}
+
 #[test]
 fn test_unsat() {
-    use crate::{ast, Config, SatResult};
-    let cfg = Config::new();
-    let ctx = Context::new(&cfg);
-    let solver = Solver::new(&ctx);
-    solver.assert(&ast::Bool::from_bool(&ctx, false));
+    use crate::SatResult;
+    let solver = Solver::new();
+    solver.assert(false);
     assert_eq!(solver.check(), SatResult::Unsat);
     assert!(solver.get_model().is_none());
 }
