@@ -6,8 +6,8 @@ use std::fmt;
 use z3_sys::*;
 
 use crate::{
-    ast::{Ast, Bool, Dynamic},
     Context, Model, Optimize, Params, SatResult, Statistics, Symbol,
+    ast::{Ast, Bool, Dynamic},
 };
 
 use num::{
@@ -15,15 +15,21 @@ use num::{
     rational::BigRational,
 };
 
-impl<'ctx> Optimize<'ctx> {
-    unsafe fn wrap(ctx: &'ctx Context, z3_opt: Z3_optimize) -> Optimize<'ctx> {
-        Z3_optimize_inc_ref(ctx.z3_ctx, z3_opt);
-        Optimize { ctx, z3_opt }
+impl Optimize {
+    unsafe fn wrap(ctx: &Context, z3_opt: Z3_optimize) -> Optimize {
+        unsafe {
+            Z3_optimize_inc_ref(ctx.z3_ctx.0, z3_opt);
+        }
+        Optimize {
+            ctx: ctx.clone(),
+            z3_opt,
+        }
     }
 
     /// Create a new optimize context.
-    pub fn new(ctx: &'ctx Context) -> Optimize<'ctx> {
-        unsafe { Self::wrap(ctx, Z3_mk_optimize(ctx.z3_ctx)) }
+    pub fn new() -> Optimize {
+        let ctx = &Context::thread_local();
+        unsafe { Self::wrap(ctx, Z3_mk_optimize(ctx.z3_ctx.0)) }
     }
 
     /// Parse an SMT-LIB2 string with assertions, soft constraints and optimization objectives.
@@ -31,13 +37,13 @@ impl<'ctx> Optimize<'ctx> {
     pub fn from_string<T: Into<Vec<u8>>>(&self, source_string: T) {
         let source_cstring = CString::new(source_string).unwrap();
         unsafe {
-            Z3_optimize_from_string(self.ctx.z3_ctx, self.z3_opt, source_cstring.as_ptr());
+            Z3_optimize_from_string(self.ctx.z3_ctx.0, self.z3_opt, source_cstring.as_ptr());
         }
     }
 
     /// Get this optimizers 's context.
-    pub fn get_context(&self) -> &'ctx Context {
-        self.ctx
+    pub fn get_context(&self) -> &Context {
+        &self.ctx
     }
 
     /// Assert hard constraint to the optimization context.
@@ -47,8 +53,8 @@ impl<'ctx> Optimize<'ctx> {
     /// - [`Optimize::assert_soft()`]
     /// - [`Optimize::maximize()`]
     /// - [`Optimize::minimize()`]
-    pub fn assert(&self, ast: &impl Ast<'ctx>) {
-        unsafe { Z3_optimize_assert(self.ctx.z3_ctx, self.z3_opt, ast.get_z3_ast()) };
+    pub fn assert(&self, ast: &impl Ast) {
+        unsafe { Z3_optimize_assert(self.ctx.z3_ctx.0, self.z3_opt, ast.get_z3_ast()) };
     }
 
     /// Assert a constraint `a` into the solver, and track it (in the
@@ -66,9 +72,11 @@ impl<'ctx> Optimize<'ctx> {
     ///
     /// - [`Optimize::assert()`]
     /// - [`Optimize::assert_soft()`]
-    pub fn assert_and_track(&self, ast: &Bool<'ctx>, p: &Bool<'ctx>) {
-        debug!("assert_and_track: {:?}", ast);
-        unsafe { Z3_optimize_assert_and_track(self.ctx.z3_ctx, self.z3_opt, ast.z3_ast, p.z3_ast) };
+    pub fn assert_and_track(&self, ast: &Bool, p: &Bool) {
+        debug!("assert_and_track: {ast:?}");
+        unsafe {
+            Z3_optimize_assert_and_track(self.ctx.z3_ctx.0, self.z3_opt, ast.z3_ast, p.z3_ast)
+        };
     }
 
     /// Assert soft constraint to the optimization context.
@@ -80,15 +88,15 @@ impl<'ctx> Optimize<'ctx> {
     /// - [`Optimize::assert()`]
     /// - [`Optimize::maximize()`]
     /// - [`Optimize::minimize()`]
-    pub fn assert_soft(&self, ast: &impl Ast<'ctx>, weight: impl Weight, group: Option<Symbol>) {
+    pub fn assert_soft(&self, ast: &impl Ast, weight: impl Weight, group: Option<Symbol>) {
         let weight_string = weight.to_string();
         let weight_cstring = CString::new(weight_string).unwrap();
         let group = group
-            .map(|g| g.as_z3_symbol(self.ctx))
+            .map(|g| g.as_z3_symbol())
             .unwrap_or_else(std::ptr::null_mut);
         unsafe {
             Z3_optimize_assert_soft(
-                self.ctx.z3_ctx,
+                self.ctx.z3_ctx.0,
                 self.z3_opt,
                 ast.get_z3_ast(),
                 weight_cstring.as_ptr(),
@@ -103,13 +111,13 @@ impl<'ctx> Optimize<'ctx> {
     ///
     /// - [`Optimize::assert()`]
     /// - [`Optimize::minimize()`]
-    pub fn maximize(&self, ast: &impl Ast<'ctx>) {
+    pub fn maximize(&self, ast: &impl Ast) {
         // https://github.com/Z3Prover/z3/blob/09f911d8a84cd91988e5b96b69485b2a9a2edba3/src/opt/opt_context.cpp#L118-L120
         assert!(matches!(
             ast.get_sort().kind(),
             SortKind::Int | SortKind::Real | SortKind::BV
         ));
-        unsafe { Z3_optimize_maximize(self.ctx.z3_ctx, self.z3_opt, ast.get_z3_ast()) };
+        unsafe { Z3_optimize_maximize(self.ctx.z3_ctx.0, self.z3_opt, ast.get_z3_ast()) };
     }
 
     /// Add a minimization constraint.
@@ -118,12 +126,12 @@ impl<'ctx> Optimize<'ctx> {
     ///
     /// - [`Optimize::assert()`]
     /// - [`Optimize::maximize()`]
-    pub fn minimize(&self, ast: &impl Ast<'ctx>) {
+    pub fn minimize(&self, ast: &impl Ast) {
         assert!(matches!(
             ast.get_sort().kind(),
             SortKind::Int | SortKind::Real | SortKind::BV
         ));
-        unsafe { Z3_optimize_minimize(self.ctx.z3_ctx, self.z3_opt, ast.get_z3_ast()) };
+        unsafe { Z3_optimize_minimize(self.ctx.z3_ctx.0, self.z3_opt, ast.get_z3_ast()) };
     }
 
     /// Return a subset of the assumptions provided to either the last
@@ -146,19 +154,19 @@ impl<'ctx> Optimize<'ctx> {
     /// # See also:
     ///
     /// - [`Optimize::check`]
-    pub fn get_unsat_core(&self) -> Vec<Bool<'ctx>> {
-        let z3_unsat_core = unsafe { Z3_optimize_get_unsat_core(self.ctx.z3_ctx, self.z3_opt) };
+    pub fn get_unsat_core(&self) -> Vec<Bool> {
+        let z3_unsat_core = unsafe { Z3_optimize_get_unsat_core(self.ctx.z3_ctx.0, self.z3_opt) };
         if z3_unsat_core.is_null() {
             return vec![];
         }
 
-        let len = unsafe { Z3_ast_vector_size(self.ctx.z3_ctx, z3_unsat_core) };
+        let len = unsafe { Z3_ast_vector_size(self.ctx.z3_ctx.0, z3_unsat_core) };
 
         let mut unsat_core = Vec::with_capacity(len as usize);
 
         for i in 0..len {
-            let elem = unsafe { Z3_ast_vector_get(self.ctx.z3_ctx, z3_unsat_core, i) };
-            let elem = unsafe { Bool::wrap(self.ctx, elem) };
+            let elem = unsafe { Z3_ast_vector_get(self.ctx.z3_ctx.0, z3_unsat_core, i) };
+            let elem = unsafe { Bool::wrap(&self.ctx, elem) };
             unsat_core.push(elem);
         }
 
@@ -175,7 +183,7 @@ impl<'ctx> Optimize<'ctx> {
     ///
     /// - [`Optimize::pop()`]
     pub fn push(&self) {
-        unsafe { Z3_optimize_push(self.ctx.z3_ctx, self.z3_opt) };
+        unsafe { Z3_optimize_push(self.ctx.z3_ctx.0, self.z3_opt) };
     }
 
     /// Backtrack one level.
@@ -189,7 +197,7 @@ impl<'ctx> Optimize<'ctx> {
     ///
     /// - [`Optimize::push()`]
     pub fn pop(&self) {
-        unsafe { Z3_optimize_pop(self.ctx.z3_ctx, self.z3_opt) };
+        unsafe { Z3_optimize_pop(self.ctx.z3_ctx.0, self.z3_opt) };
     }
 
     /// Check consistency and produce optimal values.
@@ -197,11 +205,11 @@ impl<'ctx> Optimize<'ctx> {
     /// # See also:
     ///
     /// - [`Optimize::get_model()`]
-    pub fn check(&self, assumptions: &[Bool<'ctx>]) -> SatResult {
+    pub fn check(&self, assumptions: &[Bool]) -> SatResult {
         let assumptions: Vec<Z3_ast> = assumptions.iter().map(|a| a.z3_ast).collect();
         match unsafe {
             Z3_optimize_check(
-                self.ctx.z3_ctx,
+                self.ctx.z3_ctx.0,
                 self.z3_opt,
                 assumptions.len().try_into().unwrap(),
                 assumptions.as_ptr(),
@@ -219,25 +227,25 @@ impl<'ctx> Optimize<'ctx> {
     /// The error handler is invoked if a model is not available because
     /// the commands above were not invoked for the given optimization
     /// solver, or if the result was [`SatResult::Unsat`].
-    pub fn get_model(&self) -> Option<Model<'ctx>> {
+    pub fn get_model(&self) -> Option<Model> {
         Model::of_optimize(self)
     }
 
     /// Retrieve the objectives for the last [`Optimize::check()`].
     ///
     /// This contains maximize/minimize objectives and grouped soft constraints.
-    pub fn get_objectives(&self) -> Vec<Dynamic<'ctx>> {
+    pub fn get_objectives(&self) -> Vec<Dynamic> {
         let (z3_objectives, len) = unsafe {
-            let objectives = Z3_optimize_get_objectives(self.ctx.z3_ctx, self.z3_opt);
-            let len = Z3_ast_vector_size(self.ctx.z3_ctx, objectives);
+            let objectives = Z3_optimize_get_objectives(self.ctx.z3_ctx.0, self.z3_opt);
+            let len = Z3_ast_vector_size(self.ctx.z3_ctx.0, objectives);
             (objectives, len)
         };
 
         let mut objectives = Vec::with_capacity(len as usize);
 
         for i in 0..len {
-            let elem = unsafe { Z3_ast_vector_get(self.ctx.z3_ctx, z3_objectives, i) };
-            let elem = unsafe { Dynamic::wrap(self.ctx, elem) };
+            let elem = unsafe { Z3_ast_vector_get(self.ctx.z3_ctx.0, z3_objectives, i) };
+            let elem = unsafe { Dynamic::wrap(&self.ctx, elem) };
             objectives.push(elem);
         }
 
@@ -248,7 +256,7 @@ impl<'ctx> Optimize<'ctx> {
     ///
     /// Use this method when [`Optimize::check()`] returns [`SatResult::Unknown`].
     pub fn get_reason_unknown(&self) -> Option<String> {
-        let p = unsafe { Z3_optimize_get_reason_unknown(self.ctx.z3_ctx, self.z3_opt) };
+        let p = unsafe { Z3_optimize_get_reason_unknown(self.ctx.z3_ctx.0, self.z3_opt) };
         if p.is_null() {
             return None;
         }
@@ -259,24 +267,24 @@ impl<'ctx> Optimize<'ctx> {
     }
 
     /// Configure the parameters for this Optimize.
-    pub fn set_params(&self, params: &Params<'ctx>) {
-        unsafe { Z3_optimize_set_params(self.ctx.z3_ctx, self.z3_opt, params.z3_params) };
+    pub fn set_params(&self, params: &Params) {
+        unsafe { Z3_optimize_set_params(self.ctx.z3_ctx.0, self.z3_opt, params.z3_params) };
     }
 
     /// Retrieve the statistics for the last [`Optimize::check()`].
-    pub fn get_statistics(&self) -> Statistics<'ctx> {
+    pub fn get_statistics(&self) -> Statistics {
         unsafe {
             Statistics::wrap(
-                self.ctx,
-                Z3_optimize_get_statistics(self.ctx.z3_ctx, self.z3_opt),
+                &self.ctx,
+                Z3_optimize_get_statistics(self.ctx.z3_ctx.0, self.z3_opt),
             )
         }
     }
 }
 
-impl<'ctx> fmt::Display for Optimize<'ctx> {
+impl fmt::Display for Optimize {
     fn fmt(&self, f: &mut fmt::Formatter) -> Result<(), fmt::Error> {
-        let p = unsafe { Z3_optimize_to_string(self.ctx.z3_ctx, self.z3_opt) };
+        let p = unsafe { Z3_optimize_to_string(self.ctx.z3_ctx.0, self.z3_opt) };
         if p.is_null() {
             return Result::Err(fmt::Error);
         }
@@ -287,15 +295,15 @@ impl<'ctx> fmt::Display for Optimize<'ctx> {
     }
 }
 
-impl<'ctx> fmt::Debug for Optimize<'ctx> {
+impl fmt::Debug for Optimize {
     fn fmt(&self, f: &mut fmt::Formatter) -> Result<(), fmt::Error> {
         <Self as fmt::Display>::fmt(self, f)
     }
 }
 
-impl<'ctx> Drop for Optimize<'ctx> {
+impl Drop for Optimize {
     fn drop(&mut self) {
-        unsafe { Z3_optimize_dec_ref(self.ctx.z3_ctx, self.z3_opt) };
+        unsafe { Z3_optimize_dec_ref(self.ctx.z3_ctx.0, self.z3_opt) };
     }
 }
 
