@@ -1,12 +1,11 @@
 use std::convert::TryInto;
 use std::ffi::CStr;
 use std::fmt;
-
 use z3_sys::*;
 
-use crate::{ast::Ast, Context, Pattern};
+use crate::{Context, Pattern, ast::Ast};
 
-impl<'ctx> Pattern<'ctx> {
+impl Pattern {
     /// Create a pattern for quantifier instantiation.
     ///
     /// Z3 uses pattern matching to instantiate quantifiers. If a
@@ -25,30 +24,31 @@ impl<'ctx> Pattern<'ctx> {
     ///
     /// - `ast::forall_const()`
     /// - `ast::exists_const()`
-    pub fn new(ctx: &'ctx Context, terms: &[&dyn Ast]) -> Pattern<'ctx> {
+    pub fn new(terms: &[&dyn Ast]) -> Pattern {
+        let ctx = &Context::thread_local();
         assert!(!terms.is_empty());
         assert!(terms.iter().all(|t| t.get_ctx().z3_ctx == ctx.z3_ctx));
 
         let terms: Vec<_> = terms.iter().map(|t| t.get_z3_ast()).collect();
 
         Pattern {
-            ctx,
+            ctx: ctx.clone(),
             z3_pattern: unsafe {
                 let p = Z3_mk_pattern(
-                    ctx.z3_ctx,
+                    ctx.z3_ctx.0,
                     terms.len().try_into().unwrap(),
                     terms.as_ptr() as *const Z3_ast,
                 );
-                Z3_inc_ref(ctx.z3_ctx, p as Z3_ast);
+                Z3_inc_ref(ctx.z3_ctx.0, p as Z3_ast);
                 p
             },
         }
     }
 }
 
-impl<'ctx> fmt::Debug for Pattern<'ctx> {
+impl fmt::Debug for Pattern {
     fn fmt(&self, f: &mut fmt::Formatter) -> Result<(), fmt::Error> {
-        let p = unsafe { Z3_pattern_to_string(self.ctx.z3_ctx, self.z3_pattern) };
+        let p = unsafe { Z3_pattern_to_string(self.ctx.z3_ctx.0, self.z3_pattern) };
         if p.is_null() {
             return Result::Err(fmt::Error);
         }
@@ -59,16 +59,16 @@ impl<'ctx> fmt::Debug for Pattern<'ctx> {
     }
 }
 
-impl<'ctx> fmt::Display for Pattern<'ctx> {
+impl fmt::Display for Pattern {
     fn fmt(&self, f: &mut fmt::Formatter) -> Result<(), fmt::Error> {
         <Self as fmt::Debug>::fmt(self, f)
     }
 }
 
-impl<'ctx> Drop for Pattern<'ctx> {
+impl Drop for Pattern {
     fn drop(&mut self) {
         unsafe {
-            Z3_dec_ref(self.ctx.z3_ctx, self.z3_pattern as Z3_ast);
+            Z3_dec_ref(self.ctx.z3_ctx.0, self.z3_pattern as Z3_ast);
         }
     }
 }
