@@ -2,25 +2,28 @@ use std::convert::TryInto;
 use std::ffi::CStr;
 use std::fmt;
 use std::ops::Deref;
-
 use z3_sys::*;
 
-use crate::{ast, ast::Ast, Context, FuncDecl, RecFuncDecl, Sort, Symbol};
+use crate::{Context, FuncDecl, RecFuncDecl, Sort, Symbol, ast, ast::Ast};
 
-impl<'ctx> RecFuncDecl<'ctx> {
-    pub(crate) unsafe fn wrap(ctx: &'ctx Context, z3_func_decl: Z3_func_decl) -> Self {
-        Z3_inc_ref(ctx.z3_ctx, Z3_func_decl_to_ast(ctx.z3_ctx, z3_func_decl));
-        Self { ctx, z3_func_decl }
+impl RecFuncDecl {
+    pub(crate) unsafe fn wrap(ctx: &Context, z3_func_decl: Z3_func_decl) -> Self {
+        unsafe {
+            Z3_inc_ref(
+                ctx.z3_ctx.0,
+                Z3_func_decl_to_ast(ctx.z3_ctx.0, z3_func_decl),
+            );
+        }
+        Self {
+            ctx: ctx.clone(),
+            z3_func_decl,
+        }
     }
 
-    pub fn new<S: Into<Symbol>>(
-        ctx: &'ctx Context,
-        name: S,
-        domain: &[&Sort<'ctx>],
-        range: &Sort<'ctx>,
-    ) -> Self {
+    pub fn new<S: Into<Symbol>>(name: S, domain: &[&Sort], range: &Sort) -> Self {
+        let ctx = &Context::thread_local();
         assert!(domain.iter().all(|s| s.ctx.z3_ctx == ctx.z3_ctx));
-        assert_eq!(ctx.z3_ctx, range.ctx.z3_ctx);
+        assert_eq!(ctx.z3_ctx.0, range.ctx.z3_ctx.0);
 
         let domain: Vec<_> = domain.iter().map(|s| s.z3_sort).collect();
 
@@ -28,8 +31,8 @@ impl<'ctx> RecFuncDecl<'ctx> {
             Self::wrap(
                 ctx,
                 Z3_mk_rec_func_decl(
-                    ctx.z3_ctx,
-                    name.into().as_z3_symbol(ctx),
+                    ctx.z3_ctx.0,
+                    name.into().as_z3_symbol(),
                     domain.len().try_into().unwrap(),
                     domain.as_ptr(),
                     range.z3_sort,
@@ -43,24 +46,20 @@ impl<'ctx> RecFuncDecl<'ctx> {
     /// ```
     /// # use z3::{Config, Context, RecFuncDecl, Solver, Sort, Symbol, ast::Int, SatResult};
     /// # use std::convert::TryInto;
-    /// # let cfg = Config::new();
-    /// # let ctx = Context::new(&cfg);
     /// let mut f = RecFuncDecl::new(
-    ///     &ctx,
     ///     "f",
-    ///     &[&Sort::int(&ctx)],
-    ///     &Sort::int(&ctx));
-    /// let n = Int::new_const(&ctx, "n");
+    ///     &[&Sort::int()],
+    ///     &Sort::int());
+    /// let n = Int::new_const("n");
     /// f.add_def(
     ///     &[&n],
-    ///     &Int::add(&ctx, &[&n, &Int::from_i64(&ctx, 1)])
+    ///     &Int::add(&[&n, &Int::from_i64(1)])
     /// );
     ///
     /// let f_of_n = &f.apply(&[&n.clone()]);
     ///
-    /// let solver = Solver::new(&ctx);
+    /// let solver = Solver::new();
     /// let forall: z3::ast::Bool = z3::ast::forall_const(
-    ///         &ctx,
     ///         &[&n],
     ///         &[],
     ///         &n.lt(&f_of_n.as_int().unwrap())
@@ -72,19 +71,19 @@ impl<'ctx> RecFuncDecl<'ctx> {
     /// ```
     ///
     /// Note that `args` should have the types corresponding to the `domain` of the `RecFuncDecl`.
-    pub fn add_def(&self, args: &[&dyn ast::Ast<'ctx>], body: &dyn Ast<'ctx>) {
+    pub fn add_def(&self, args: &[&dyn ast::Ast], body: &dyn Ast) {
         assert!(args.iter().all(|arg| arg.get_ctx() == body.get_ctx()));
-        assert_eq!(self.ctx, body.get_ctx());
+        assert_eq!(&self.ctx, body.get_ctx());
 
         let mut args: Vec<_> = args.iter().map(|s| s.get_z3_ast()).collect();
         unsafe {
             assert_eq!(
                 body.get_sort().z3_sort,
-                Z3_get_range(self.ctx.z3_ctx, self.z3_func_decl)
+                Z3_get_range(self.ctx.z3_ctx.0, self.z3_func_decl)
             );
 
             Z3_add_rec_def(
-                self.ctx.z3_ctx,
+                self.ctx.z3_ctx.0,
                 self.z3_func_decl,
                 self.arity() as u32,
                 args.as_mut_ptr(),
@@ -94,9 +93,9 @@ impl<'ctx> RecFuncDecl<'ctx> {
     }
 }
 
-impl<'ctx> fmt::Display for RecFuncDecl<'ctx> {
+impl fmt::Display for RecFuncDecl {
     fn fmt(&self, f: &mut fmt::Formatter) -> Result<(), fmt::Error> {
-        let p = unsafe { Z3_func_decl_to_string(self.ctx.z3_ctx, self.z3_func_decl) };
+        let p = unsafe { Z3_func_decl_to_string(self.ctx.z3_ctx.0, self.z3_func_decl) };
         if p.is_null() {
             return Result::Err(fmt::Error);
         }
@@ -107,25 +106,25 @@ impl<'ctx> fmt::Display for RecFuncDecl<'ctx> {
     }
 }
 
-impl<'ctx> fmt::Debug for RecFuncDecl<'ctx> {
+impl fmt::Debug for RecFuncDecl {
     fn fmt(&self, f: &mut fmt::Formatter) -> Result<(), fmt::Error> {
         <Self as fmt::Display>::fmt(self, f)
     }
 }
 
-impl<'ctx> Drop for RecFuncDecl<'ctx> {
+impl Drop for RecFuncDecl {
     fn drop(&mut self) {
         unsafe {
             Z3_dec_ref(
-                self.ctx.z3_ctx,
-                Z3_func_decl_to_ast(self.ctx.z3_ctx, self.z3_func_decl),
+                self.ctx.z3_ctx.0,
+                Z3_func_decl_to_ast(self.ctx.z3_ctx.0, self.z3_func_decl),
             );
         }
     }
 }
 
-impl<'ctx> Deref for RecFuncDecl<'ctx> {
-    type Target = FuncDecl<'ctx>;
+impl Deref for RecFuncDecl {
+    type Target = FuncDecl;
 
     fn deref(&self) -> &Self::Target {
         unsafe { &*(self as *const _ as *const Self::Target) }
