@@ -28,11 +28,16 @@ pub struct StatisticsEntry {
     pub value: StatisticsValue,
 }
 
-impl<'ctx> Statistics<'ctx> {
+impl Statistics {
     /// Wrap a raw [`Z3_stats`], managing refcounts.
-    pub(crate) unsafe fn wrap(ctx: &'ctx Context, z3_stats: Z3_stats) -> Statistics<'ctx> {
-        Z3_stats_inc_ref(ctx.z3_ctx, z3_stats);
-        Statistics { ctx, z3_stats }
+    pub(crate) unsafe fn wrap(ctx: &Context, z3_stats: Z3_stats) -> Statistics {
+        unsafe {
+            Z3_stats_inc_ref(ctx.z3_ctx.0, z3_stats);
+        }
+        Statistics {
+            ctx: ctx.clone(),
+            z3_stats,
+        }
     }
 
     /// Get the statistics value at the given index.
@@ -41,23 +46,29 @@ impl<'ctx> Statistics<'ctx> {
     ///
     /// This assumes that `idx` is a valid index.
     unsafe fn value_at_idx(&self, idx: u32) -> StatisticsValue {
-        if Z3_stats_is_uint(self.ctx.z3_ctx, self.z3_stats, idx) {
-            StatisticsValue::UInt(Z3_stats_get_uint_value(self.ctx.z3_ctx, self.z3_stats, idx))
-        } else {
-            StatisticsValue::Double(Z3_stats_get_double_value(
-                self.ctx.z3_ctx,
-                self.z3_stats,
-                idx,
-            ))
+        unsafe {
+            if Z3_stats_is_uint(self.ctx.z3_ctx.0, self.z3_stats, idx) {
+                StatisticsValue::UInt(Z3_stats_get_uint_value(
+                    self.ctx.z3_ctx.0,
+                    self.z3_stats,
+                    idx,
+                ))
+            } else {
+                StatisticsValue::Double(Z3_stats_get_double_value(
+                    self.ctx.z3_ctx.0,
+                    self.z3_stats,
+                    idx,
+                ))
+            }
         }
     }
 
     /// Get the statistics value for the given `key`.
     pub fn value(&self, key: &str) -> Option<StatisticsValue> {
         unsafe {
-            let size = Z3_stats_size(self.ctx.z3_ctx, self.z3_stats);
+            let size = Z3_stats_size(self.ctx.z3_ctx.0, self.z3_stats);
             for idx in 0..size {
-                let k = CStr::from_ptr(Z3_stats_get_key(self.ctx.z3_ctx, self.z3_stats, idx));
+                let k = CStr::from_ptr(Z3_stats_get_key(self.ctx.z3_ctx.0, self.z3_stats, idx));
                 if k.to_str().unwrap() == key {
                     return Some(self.value_at_idx(idx));
                 }
@@ -68,9 +79,9 @@ impl<'ctx> Statistics<'ctx> {
 
     /// Iterate over all of the entries in this set of statistics.
     pub fn entries(&self) -> impl Iterator<Item = StatisticsEntry> + '_ {
-        let p = unsafe { Z3_stats_size(self.ctx.z3_ctx, self.z3_stats) };
+        let p = unsafe { Z3_stats_size(self.ctx.z3_ctx.0, self.z3_stats) };
         (0..p).map(move |n| unsafe {
-            let t = Z3_stats_get_key(self.ctx.z3_ctx, self.z3_stats, n);
+            let t = Z3_stats_get_key(self.ctx.z3_ctx.0, self.z3_stats, n);
             StatisticsEntry {
                 key: CStr::from_ptr(t).to_str().unwrap().to_owned(),
                 value: self.value_at_idx(n),
@@ -79,19 +90,19 @@ impl<'ctx> Statistics<'ctx> {
     }
 }
 
-impl<'ctx> Clone for Statistics<'ctx> {
+impl Clone for Statistics {
     fn clone(&self) -> Self {
-        unsafe { Self::wrap(self.ctx, self.z3_stats) }
+        unsafe { Self::wrap(&self.ctx, self.z3_stats) }
     }
 }
 
-impl<'ctx> fmt::Display for Statistics<'ctx> {
+impl fmt::Display for Statistics {
     fn fmt(&self, f: &mut fmt::Formatter) -> Result<(), fmt::Error> {
         write!(f, "<z3.stats>")
     }
 }
 
-impl<'ctx> fmt::Debug for Statistics<'ctx> {
+impl fmt::Debug for Statistics {
     fn fmt(&self, f: &mut fmt::Formatter) -> Result<(), fmt::Error> {
         let mut s = f.debug_struct("Statistics");
         for e in self.entries() {
@@ -104,10 +115,10 @@ impl<'ctx> fmt::Debug for Statistics<'ctx> {
     }
 }
 
-impl<'ctx> Drop for Statistics<'ctx> {
+impl Drop for Statistics {
     fn drop(&mut self) {
         unsafe {
-            Z3_stats_dec_ref(self.ctx.z3_ctx, self.z3_stats);
+            Z3_stats_dec_ref(self.ctx.z3_ctx.0, self.z3_stats);
         }
     }
 }
