@@ -5,41 +5,42 @@ use std::os::raw::c_uint;
 use std::result::Result;
 use std::str::Utf8Error;
 use std::time::Duration;
-
 use z3_sys::*;
 
 use crate::{ApplyResult, Context, Goal, Params, Probe, Solver, Tactic};
 
-impl<'ctx> ApplyResult<'ctx> {
-    unsafe fn wrap(ctx: &'ctx Context, z3_apply_result: Z3_apply_result) -> ApplyResult<'ctx> {
-        Z3_apply_result_inc_ref(ctx.z3_ctx, z3_apply_result);
+impl ApplyResult {
+    unsafe fn wrap(ctx: &Context, z3_apply_result: Z3_apply_result) -> ApplyResult {
+        unsafe {
+            Z3_apply_result_inc_ref(ctx.z3_ctx.0, z3_apply_result);
+        }
         ApplyResult {
-            ctx,
+            ctx: ctx.clone(),
             z3_apply_result,
         }
     }
 
-    pub fn list_subgoals(self) -> impl Iterator<Item = Goal<'ctx>> {
+    pub fn list_subgoals(self) -> impl Iterator<Item = Goal> {
         let num_subgoals =
-            unsafe { Z3_apply_result_get_num_subgoals(self.ctx.z3_ctx, self.z3_apply_result) };
+            unsafe { Z3_apply_result_get_num_subgoals(self.ctx.z3_ctx.0, self.z3_apply_result) };
         (0..num_subgoals).map(move |i| unsafe {
             Goal::wrap(
-                self.ctx,
-                Z3_apply_result_get_subgoal(self.ctx.z3_ctx, self.z3_apply_result, i),
+                &self.ctx,
+                Z3_apply_result_get_subgoal(self.ctx.z3_ctx.0, self.z3_apply_result, i),
             )
         })
     }
 }
 
-impl<'ctx> Drop for ApplyResult<'ctx> {
+impl Drop for ApplyResult {
     fn drop(&mut self) {
         unsafe {
-            Z3_apply_result_dec_ref(self.ctx.z3_ctx, self.z3_apply_result);
+            Z3_apply_result_dec_ref(self.ctx.z3_ctx.0, self.z3_apply_result);
         }
     }
 }
 
-impl<'ctx> Tactic<'ctx> {
+impl Tactic {
     /// Iterate through the valid tactic names.
     ///
     /// # Example
@@ -49,22 +50,28 @@ impl<'ctx> Tactic<'ctx> {
     ///
     /// let cfg = Config::new();
     /// let ctx = Context::new(&cfg);
-    /// let tactics: Vec<_> = Tactic::list_all(&ctx).filter_map(|r| r.ok()).collect();
-    /// assert!(tactics.contains(&"ufbv"));
+    /// let tactics: Vec<_> = Tactic::list_all().into_iter().filter_map(|r| r.ok()).collect();
+    /// assert!(tactics.contains(&"ufbv".to_string()));
     /// ```
-    pub fn list_all(
-        ctx: &'ctx Context,
-    ) -> impl Iterator<Item = std::result::Result<&'ctx str, Utf8Error>> {
-        let p = unsafe { Z3_get_num_tactics(ctx.z3_ctx) };
-        (0..p).map(move |n| {
-            let t = unsafe { Z3_get_tactic_name(ctx.z3_ctx, n) };
-            unsafe { CStr::from_ptr(t) }.to_str()
-        })
+    pub fn list_all() -> Vec<Result<String, Utf8Error>> {
+        let ctx = &Context::thread_local();
+        let p = unsafe { Z3_get_num_tactics(ctx.z3_ctx.0) };
+        (0..p)
+            .map(move |n| {
+                let t = unsafe { Z3_get_tactic_name(ctx.z3_ctx.0, n) };
+                unsafe { CStr::from_ptr(t) }.to_str().map(String::from)
+            })
+            .collect()
     }
 
-    unsafe fn wrap(ctx: &'ctx Context, z3_tactic: Z3_tactic) -> Tactic<'ctx> {
-        Z3_tactic_inc_ref(ctx.z3_ctx, z3_tactic);
-        Tactic { ctx, z3_tactic }
+    unsafe fn wrap(ctx: &Context, z3_tactic: Z3_tactic) -> Tactic {
+        unsafe {
+            Z3_tactic_inc_ref(ctx.z3_ctx.0, z3_tactic);
+        }
+        Tactic {
+            ctx: ctx.clone(),
+            z3_tactic,
+        }
     }
 
     /// Create a tactic by name.
@@ -76,136 +83,141 @@ impl<'ctx> Tactic<'ctx> {
     ///
     /// let cfg = Config::new();
     /// let ctx = Context::new(&cfg);
-    /// let tactic = Tactic::new(&ctx, "nlsat");
+    /// let tactic = Tactic::new("nlsat");
     /// ```
     ///
     /// # See also
     ///
     /// - [`Tactic::list_all()`]
-    pub fn new(ctx: &'ctx Context, name: &str) -> Tactic<'ctx> {
+    pub fn new(name: &str) -> Tactic {
+        let ctx = &Context::thread_local();
         let tactic_name = CString::new(name).unwrap();
-        unsafe { Self::wrap(ctx, Z3_mk_tactic(ctx.z3_ctx, tactic_name.as_ptr())) }
+
+        unsafe {
+            let tactic = Z3_mk_tactic(ctx.z3_ctx.0, tactic_name.as_ptr());
+            if tactic.is_null() {
+                panic!("{name} is an invalid tactic");
+            } else {
+                Self::wrap(ctx, tactic)
+            }
+        }
     }
 
     /// Return a tactic that just return the given goal.
-    pub fn create_skip(ctx: &'ctx Context) -> Tactic<'ctx> {
-        unsafe { Self::wrap(ctx, Z3_tactic_skip(ctx.z3_ctx)) }
+    pub fn create_skip() -> Tactic {
+        let ctx = &Context::thread_local();
+        unsafe { Self::wrap(ctx, Z3_tactic_skip(ctx.z3_ctx.0)) }
     }
 
     /// Return a tactic that always fails.
-    pub fn create_fail(ctx: &'ctx Context) -> Tactic<'ctx> {
-        unsafe { Self::wrap(ctx, Z3_tactic_fail(ctx.z3_ctx)) }
+    pub fn create_fail() -> Tactic {
+        let ctx = &Context::thread_local();
+        unsafe { Self::wrap(ctx, Z3_tactic_fail(ctx.z3_ctx.0)) }
     }
 
     /// Return a tactic that keeps applying `t` until the goal is not modified anymore or the maximum
     /// number of iterations `max` is reached.
-    pub fn repeat(ctx: &'ctx Context, t: &Tactic<'ctx>, max: u32) -> Tactic<'ctx> {
-        unsafe { Self::wrap(ctx, Z3_tactic_repeat(ctx.z3_ctx, t.z3_tactic, max)) }
+    pub fn repeat(t: &Tactic, max: u32) -> Tactic {
+        let ctx = &Context::thread_local();
+        unsafe { Self::wrap(ctx, Z3_tactic_repeat(ctx.z3_ctx.0, t.z3_tactic, max)) }
     }
 
     /// Return a tactic that applies the current tactic to a given goal, failing
     /// if it doesn't terminate within the period specified by `timeout`.
-    pub fn try_for(&self, timeout: Duration) -> Tactic<'ctx> {
+    pub fn try_for(&self, timeout: Duration) -> Tactic {
         let timeout_ms = c_uint::try_from(timeout.as_millis()).unwrap_or(c_uint::MAX);
         unsafe {
             Self::wrap(
-                self.ctx,
-                Z3_tactic_try_for(self.ctx.z3_ctx, self.z3_tactic, timeout_ms),
+                &self.ctx,
+                Z3_tactic_try_for(self.ctx.z3_ctx.0, self.z3_tactic, timeout_ms),
             )
         }
     }
 
     /// Return a tactic that applies the current tactic to a given goal and
     /// the `then_tactic` to every subgoal produced by the original tactic.
-    pub fn and_then(&self, then_tactic: &Tactic<'ctx>) -> Tactic<'ctx> {
+    pub fn and_then(&self, then_tactic: &Tactic) -> Tactic {
         unsafe {
             Self::wrap(
-                self.ctx,
-                Z3_tactic_and_then(self.ctx.z3_ctx, self.z3_tactic, then_tactic.z3_tactic),
+                &self.ctx,
+                Z3_tactic_and_then(self.ctx.z3_ctx.0, self.z3_tactic, then_tactic.z3_tactic),
             )
         }
     }
 
     /// Return a tactic that current tactic to a given goal,
     /// if it fails then returns the result of `else_tactic` applied to the given goal.
-    pub fn or_else(&self, else_tactic: &Tactic<'ctx>) -> Tactic<'ctx> {
+    pub fn or_else(&self, else_tactic: &Tactic) -> Tactic {
         unsafe {
             Self::wrap(
-                self.ctx,
-                Z3_tactic_or_else(self.ctx.z3_ctx, self.z3_tactic, else_tactic.z3_tactic),
+                &self.ctx,
+                Z3_tactic_or_else(self.ctx.z3_ctx.0, self.z3_tactic, else_tactic.z3_tactic),
             )
         }
     }
 
     /// Return a tactic that applies self to a given goal if the probe `p` evaluates to true,
     /// and `t` if `p` evaluates to false.
-    pub fn probe_or_else(&self, p: &Probe<'ctx>, t: &Tactic<'ctx>) -> Tactic<'ctx> {
+    pub fn probe_or_else(&self, p: &Probe, t: &Tactic) -> Tactic {
         unsafe {
             Self::wrap(
-                self.ctx,
-                Z3_tactic_cond(self.ctx.z3_ctx, p.z3_probe, self.z3_tactic, t.z3_tactic),
+                &self.ctx,
+                Z3_tactic_cond(self.ctx.z3_ctx.0, p.z3_probe, self.z3_tactic, t.z3_tactic),
             )
         }
     }
 
     /// Return a tactic that applies itself to a given goal if the probe `p` evaluates to true.
     /// If `p` evaluates to false, then the new tactic behaves like the skip tactic.
-    pub fn when(&self, p: &Probe<'ctx>) -> Tactic<'ctx> {
+    pub fn when(&self, p: &Probe) -> Tactic {
         unsafe {
             Self::wrap(
-                self.ctx,
-                Z3_tactic_when(self.ctx.z3_ctx, p.z3_probe, self.z3_tactic),
+                &self.ctx,
+                Z3_tactic_when(self.ctx.z3_ctx.0, p.z3_probe, self.z3_tactic),
             )
         }
     }
 
     /// Return a tactic that applies `t1` to a given goal if the probe `p` evaluates to true,
     /// and `t2` if `p` evaluates to false.
-    pub fn cond(
-        ctx: &'ctx Context,
-        p: &Probe<'ctx>,
-        t1: &Tactic<'ctx>,
-        t2: &Tactic<'ctx>,
-    ) -> Tactic<'ctx> {
+    pub fn cond(p: &Probe, t1: &Tactic, t2: &Tactic) -> Tactic {
+        assert_eq!(p.ctx, t1.ctx);
+        assert_eq!(t1.ctx, t2.ctx);
         unsafe {
             Self::wrap(
-                ctx,
-                Z3_tactic_cond(ctx.z3_ctx, p.z3_probe, t1.z3_tactic, t2.z3_tactic),
+                &p.ctx,
+                Z3_tactic_cond(p.ctx.z3_ctx.0, p.z3_probe, t1.z3_tactic, t2.z3_tactic),
             )
         }
     }
 
     /// Return a tactic that fails if the probe `p` evaluates to false.
-    pub fn fail_if(ctx: &'ctx Context, p: &Probe<'ctx>) -> Tactic<'ctx> {
-        unsafe { Self::wrap(ctx, Z3_tactic_fail_if(ctx.z3_ctx, p.z3_probe)) }
+    pub fn fail_if(p: &Probe) -> Tactic {
+        unsafe { Self::wrap(&p.ctx, Z3_tactic_fail_if(p.ctx.z3_ctx.0, p.z3_probe)) }
     }
 
     /// Attempts to apply the tactic to `goal`. If the tactic succeeds, returns
     /// `Ok(_)` with a `ApplyResult`. If the tactic fails, returns `Err(_)` with
     /// an error message describing why.
-    pub fn apply(
-        &self,
-        goal: &Goal<'ctx>,
-        params: Option<&Params<'ctx>>,
-    ) -> Result<ApplyResult<'ctx>, String> {
+    pub fn apply(&self, goal: &Goal, params: Option<&Params>) -> Result<ApplyResult, String> {
         unsafe {
             let z3_apply_result = match params {
-                None => Z3_tactic_apply(self.ctx.z3_ctx, self.z3_tactic, goal.z3_goal),
+                None => Z3_tactic_apply(self.ctx.z3_ctx.0, self.z3_tactic, goal.z3_goal),
                 Some(params) => Z3_tactic_apply_ex(
-                    self.ctx.z3_ctx,
+                    self.ctx.z3_ctx.0,
                     self.z3_tactic,
                     goal.z3_goal,
                     params.z3_params,
                 ),
             };
             if z3_apply_result.is_null() {
-                let code = Z3_get_error_code(self.ctx.z3_ctx);
-                let msg = Z3_get_error_msg(self.ctx.z3_ctx, code);
+                let code = Z3_get_error_code(self.ctx.z3_ctx.0);
+                let msg = Z3_get_error_msg(self.ctx.z3_ctx.0, code);
                 Err(String::from(CStr::from_ptr(msg).to_str().unwrap_or(
                     "Couldn't retrieve error message from z3: got invalid UTF-8",
                 )))
             } else {
-                Ok(ApplyResult::wrap(self.ctx, z3_apply_result))
+                Ok(ApplyResult::wrap(&self.ctx, z3_apply_result))
             }
         }
     }
@@ -219,29 +231,29 @@ impl<'ctx> Tactic<'ctx> {
     ///
     /// let cfg = Config::new();
     /// let ctx = Context::new(&cfg);
-    /// let tactic = Tactic::new(&ctx, "qfnra");
+    /// let tactic = Tactic::new("qfnra");
     /// let solver = tactic.solver();
     ///
-    /// let x = ast::Int::new_const(&ctx, "x");
-    /// let y = ast::Int::new_const(&ctx, "y");
+    /// let x = ast::Int::new_const("x");
+    /// let y = ast::Int::new_const("y");
     ///
     /// solver.assert(&x.gt(&y));
     /// assert_eq!(solver.check(), SatResult::Sat);
     /// ```
     ///
-    pub fn solver(&self) -> Solver<'ctx> {
+    pub fn solver(&self) -> Solver {
         unsafe {
             Solver::wrap(
-                self.ctx,
-                Z3_mk_solver_from_tactic(self.ctx.z3_ctx, self.z3_tactic),
+                &self.ctx,
+                Z3_mk_solver_from_tactic(self.ctx.z3_ctx.0, self.z3_tactic),
             )
         }
     }
 }
 
-impl<'ctx> fmt::Display for Tactic<'ctx> {
+impl fmt::Display for Tactic {
     fn fmt(&self, f: &mut fmt::Formatter) -> Result<(), fmt::Error> {
-        let p = unsafe { Z3_tactic_get_help(self.ctx.z3_ctx, self.z3_tactic) };
+        let p = unsafe { Z3_tactic_get_help(self.ctx.z3_ctx.0, self.z3_tactic) };
         if p.is_null() {
             return Result::Err(fmt::Error);
         }
@@ -252,16 +264,16 @@ impl<'ctx> fmt::Display for Tactic<'ctx> {
     }
 }
 
-impl<'ctx> fmt::Debug for Tactic<'ctx> {
+impl fmt::Debug for Tactic {
     fn fmt(&self, f: &mut fmt::Formatter) -> Result<(), fmt::Error> {
         <Self as fmt::Display>::fmt(self, f)
     }
 }
 
-impl<'ctx> Drop for Tactic<'ctx> {
+impl Drop for Tactic {
     fn drop(&mut self) {
         unsafe {
-            Z3_tactic_dec_ref(self.ctx.z3_ctx, self.z3_tactic);
+            Z3_tactic_dec_ref(self.ctx.z3_ctx.0, self.z3_tactic);
         }
     }
 }
